% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-visualize.R
\name{add_contour}
\alias{add_contour}
\title{Add contours on a map}
\usage{
add_contour(
  map.layer,
  spat.raster,
  color = rev(colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100)),
  nlevels = 10,
  show.labels = FALSE,
  label.size = 4
)
}
\arguments{
\item{map.layer}{A ggplot2 object of geographic map.}

\item{spat.raster}{A SpatRaster object to be visualized.}

\item{color}{Color of contours. Default is `rev(colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100))`.}

\item{nlevels}{nlevels Default is `10`.}

\item{show.labels}{Whether to show labels on contours? Default is `FALSE`.}

\item{label.size}{label.size Default is `4`.}
}
\value{
A ggplot2 object.
}
\description{
This function is implemented to add contours on a map. It can accept any `SpatRaster` as an input.
}
\examples{
# Read the map downloaded from DataV.GeoAtlas, and create a standard dataset
data(qtp)
showtext::showtext_auto(enable = TRUE)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%<>\% rarefy_count_table()
dataset.dts \%<>\% tidy_dataset()
dataset.dts \%<>\% calc_alpha_div(measures = c("observed", "shannon"))
dataset.dts \%>\% show_dataset()

# Perform kriging interpolation for Shannon diversity indices
kri.rst <- interp_kri(map = dataset.dts$map, met = dataset.dts$met,
                      dat = dataset.dts$div$alpha, var = 'shannon', model = 'Mat', trim.dup = TRUE)

# Visualize the interpolated results
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_spatraster(spat.raster = kri.rst) \%>\%
    add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = "X.CENTER", lat.var = "Y.CENTER")
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_spatraster(spat.raster = kri.rst, show.border = FALSE) \%>\%
    add_contour(spat.raster = kri.rst, color = 'black')
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_spatraster(spat.raster = kri.rst, show.border = FALSE) \%>\%
    add_contour(spat.raster = kri.rst, color = 'black', nlevels = 10, show.labels = TRUE)
dataset.dts$map \%>\% plot_bmap() \%>\% add_contour(spat.raster = kri.rst, nlevels = 80)
dataset.dts$map \%>\% plot_bmap() \%>\% add_contour(spat.raster = kri.rst, nlevels = 10, show.labels = TRUE)
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:rarefy_count_table]{microgeo::rarefy_count_table()}}
\code{\link[microgeo:tidy_dataset]{microgeo::tidy_dataset()}}
\code{\link[microgeo:calc_alpha_div]{microgeo::calc_alpha_div()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:interp_kri]{microgeo::interp_kri()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_spatraster]{microgeo::add_spatraster()}}
\code{\link[microgeo:add_label]{microgeo::add_label()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
