% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-visualize.R
\name{add_crs}
\alias{add_crs}
\title{Add coordinate reference system to a map before the final visualization}
\usage{
add_crs(map.layer, crs = "+proj=laea +lat_0=40 +lon_0=104", expand = TRUE, ...)
}
\arguments{
\item{map.layer}{A ggplot2 object of geographic map.}

\item{crs}{Coordinate reference system (CRS) into which all data should be projected before plotting. If it is `NULL`,
we will use the CRS defined in the first `sf` layer of the plot. Default is `+proj=laea +lat_0=40 +lon_0=104`.}

\item{expand}{If `TRUE`, we will add a small expansion factor to the limits to ensure that data and axes don't overlap.
If `FALSE`, limits are taken exactly from the data or xlim/ylim. Default is `TRUE`.}

\item{...}{Parameters parsed by \code{ggplot2::coord_sf()}.}
}
\value{
A ggplot2 object.
}
\description{
This function is designed to add coordinate reference system to a map before final visualizations through
using the \code{ggplot2::coord_sf()}.
}
\examples{
data(common.map.mean)
common.map.mean  \%>\% plot_bmap()
common.map.mean  \%>\% plot_bmap() \%>\% add_north_arrow()
common.map.mean  \%>\% plot_bmap() \%>\% add_north_arrow() \%>\% add_scale_bar()
common.map.mean  \%>\% plot_bmap() \%>\% add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()
common.map.mean  \%>\% plot_bmap() \%>\% add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs(expand = FALSE)
}
\seealso{
\code{\link[ggplot2:coord_sf]{ggplot2::coord_sf()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
