% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-visualize.R
\name{add_label}
\alias{add_label}
\title{Add labels on a map}
\usage{
add_label(
  map.layer,
  dat,
  lab.var,
  lab.col = "black",
  lon.var = "longitude",
  lat.var = "latitude",
  remove.na = FALSE,
  size = 4,
  digit = 1,
  hjust = 0.5,
  vjust = 0.5
)
}
\arguments{
\item{map.layer}{A ggplot2 object of geographic map.}

\item{dat}{A `data.frame` containing the coordinates and the variable to be added on map.}

\item{lab.var}{Variable of labels to be shown (It should be included in the <dat>).}

\item{lab.col}{Color of label. Default is `black`.}

\item{lon.var}{Variable of longitude (It should be included in the <dat>). Default is `longitude`.}

\item{lat.var}{Variable of latitude (It should be included in the <dat>). Default is `latitude`.}

\item{remove.na}{Whether to remove those rows with `NA` in the <dat>? Default is `FALSE`.}

\item{size}{Text size of labels. Default is `4`.}

\item{digit}{Digit if the <lab.var> is numeric. Default is `1`.}

\item{hjust}{Adjust of labels at horizontal direction. Default is `0.5`.}

\item{vjust}{Adjust of labels at vertical direction. Default is `0.5`.}
}
\value{
A ggplot2 object.
}
\description{
This function is implemented to add label on a map. You can add any characters or numbers on the centroid
of polygons by using such a function.
}
\examples{
##### Example 1: a map with polygons named by administrative district names #####
# Initialization of the visualizations
data(common.map.mean)
class(common.map.mean)
names(common.map.mean)
showtext::showtext_auto(enable = TRUE)

# Shows administrative district names in each polygon
common.map.mean \%>\% plot_bmap() \%>\%
    add_label(dat = common.map.mean@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER')

# Shows both the administrative district names and sample numbers in each polygon
common.map.mean \%>\% plot_bmap() \%>\%
    add_label(dat = common.map.mean@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
    add_label(dat = common.map.mean@data, lab.var = 'sample.num', lon.var = 'X.CENTER', lat.var = 'Y.CENTER',
              lab.col = 'blue', vjust = 3)

# Shows both the administrative district names and average Shannon-Wiener diversity index in each polygon
common.map.mean \%>\% plot_bmap() \%>\%
    add_label(dat = common.map.mean@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
    add_label(dat = common.map.mean@data, lab.var = 'shannon_mean', lon.var = 'X.CENTER', lat.var = 'Y.CENTER',
              lab.col = 'blue', vjust = 3, digit = 4)

# Shows both the administrative district names and average Shannon-Wiener diversity index in each polygon
# Meanwhile, the map areas are also filled by colors based on the average Shannon-Wiener diversity index
common.map.mean \%>\% plot_bmap(fill = 'auto', var = 'shannon_mean') \%>\%
    add_label(dat = common.map.mean@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
    add_label(dat = common.map.mean@data, lab.var = 'shannon_mean', lon.var = 'X.CENTER', lat.var = 'Y.CENTER',
              lab.col = 'white', vjust = 3, digit = 4)

# Shows both the administrative district names and Shannon-Wiener diversity index (mean ± se) in each polygon
# Meanwhile, the map areas are also filled by colors based on the average Shannon-Wiener diversity index
common.map.mean@data$label.new <- paste0(common.map.mean@data$shannon_mean \%>\% round(., 3), " ± ",
                                         common.map.mean@data$shannon_se \%>\% round(., 3))
common.map.mean \%>\% plot_bmap(fill = 'auto', var = 'shannon_mean') \%>\%
    add_label(dat = common.map.mean@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
    add_label(dat = common.map.mean@data, lab.var = 'label.new', lon.var = 'X.CENTER', lat.var = 'Y.CENTER',
              lab.col = 'white', vjust = 3)

##### Example 2: a map with polygons named by grids #####
# Initialization of the visualizations
data(gridded.map.mean)
class(gridded.map.mean)
names(gridded.map.mean)
showtext::showtext_auto(enable = TRUE)

# Shows all grid names in each grid
gridded.map.mean \%>\% plot_bmap() \%>\%
    add_label(dat = gridded.map.mean@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER')

# Shows grid names with sampling sites in each grid
gridded.map.mean \%>\% plot_bmap() \%>\%
    add_label(dat = gridded.map.mean@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER',
              remove.na = TRUE)

# Shows sample numbers in each grid
gridded.map.mean \%>\% plot_bmap() \%>\%
    add_label(dat = gridded.map.mean@data, lab.var = 'sample.num', lon.var = 'X.CENTER', lat.var = 'Y.CENTER',
              remove.na = TRUE)

# Shows average Shannon-Wiener diversity index in each grid
gridded.map.mean \%>\% plot_bmap() \%>\%
    add_label(dat = gridded.map.mean@data, lab.var = 'shannon_mean', lon.var = 'X.CENTER', lat.var = 'Y.CENTER',
              remove.na = TRUE)

# Shows average Shannon-Wiener diversity index in each grid
# Meanwhile, the map grid are also filled by colors based on the average Shannon-Wiener diversity index
gridded.map.mean \%>\% plot_bmap(fill = 'auto', var = 'shannon_mean') \%>\%
    add_label(dat = gridded.map.mean@data, lab.var = 'shannon_mean', lon.var = 'X.CENTER', lat.var = 'Y.CENTER',
              remove.na = TRUE)

# Shows average Shannon-Wiener diversity index in each grid
# Meanwhile, the map grid are also filled by colors based on the standard error of Shannon-Wiener diversity index
gridded.map.mean \%>\% plot_bmap(fill = 'auto', var = 'shannon_se') \%>\%
    add_label(dat = gridded.map.mean@data, lab.var = 'shannon_mean', lon.var = 'X.CENTER', lat.var = 'Y.CENTER',
              remove.na = TRUE)

# Shows average Shannon-Wiener diversity index in each grid
# Meanwhile, the map grid are also filled by colors based on sample numbers
gridded.map.mean \%>\% plot_bmap(fill = 'auto', var = 'sample.num') \%>\%
    add_label(dat = gridded.map.mean@data, lab.var = 'shannon_mean', lon.var = 'X.CENTER', lat.var = 'Y.CENTER',
              remove.na = TRUE)

# Shows average Shannon-Wiener diversity index (mean ± se) in each grid
# Meanwhile, the map grid are also filled by colors based on the average Shannon-Wiener diversity index
gridded.map.mean@data$label.new <- paste0(gridded.map.mean@data$shannon_mean \%>\% round(., 1), " ± ",
                                          gridded.map.mean@data$shannon_se \%>\% round(., 1))
gridded.map.mean \%>\% plot_bmap(fill = 'auto', var = 'shannon_mean') \%>\%
    add_label(dat = gridded.map.mean@data, lab.var = 'label.new', lon.var = 'X.CENTER', lat.var = 'Y.CENTER',
              remove.na = TRUE)
}
\seealso{
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
