% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-visualize.R
\name{add_north_arrow}
\alias{add_north_arrow}
\title{Add north arrow on a map}
\usage{
add_north_arrow(
  map.layer,
  position = c("bl", "br", "tl", "tr"),
  which.north = "true",
  height = 2.5,
  width = 2.5,
  style = "auto",
  ...
)
}
\arguments{
\item{map.layer}{A ggplot2 object of geographic map.}

\item{position}{Position of the north arrow. Select one from `'bl', 'br', 'tl', 'tr'`. Default is `tr`.}

\item{which.north}{The `grid` results in a north arrow always pointing up, while the `true` always points to the north
pole from which ever corner of the map the north arrow is in. Default is `'true'`}

\item{height}{Height of north arrow. Default is `2.5` cm.}

\item{width}{Width of north arrow. Default is `2.5` cm.}

\item{style}{Style of north arrow. See \code{ggspatial::north_arrow_orienteering()} for more detailed options. Default
is `auto`, which represents the style of `north_arrow_fancy_orienteering` with the following options: `line_width = 1`,
`line_col = "black"`, `fill = c("white", "black")`, `text_col = "black"`, `text_family = ""`, `text_face = "bold"`,
`text_size = 18`, `text_angle = 0`.}

\item{...}{Parameters parsed by \code{ggspatial::annotation_north_arrow()}.}
}
\value{
A ggplot2 object.
}
\description{
This function is used to add a north arrow on a map using the \code{ggspatial::annotation_north_arrow()}.
}
\examples{
data(common.map.mean)
common.map.mean  \%>\% plot_bmap()
common.map.mean  \%>\% plot_bmap() \%>\% add_north_arrow()
}
\seealso{
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[ggspatial:north_arrow_fancy_orienteering]{ggspatial::north_arrow_fancy_orienteering()}}
\code{\link[ggspatial:annotation_north_arrow]{ggspatial::annotation_north_arrow()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
