% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-visualize.R
\name{add_point}
\alias{add_point}
\title{Add points on a map}
\usage{
add_point(
  map.layer,
  dat,
  lab.var,
  lon.var = "longitude",
  lat.var = "latitude",
  remove.na = FALSE,
  fill = NA,
  color = "gray30",
  alpha = 0.5,
  hjust = 0.5,
  vjust = 0.5
)
}
\arguments{
\item{map.layer}{A ggplot2 object of geographic map.}

\item{dat}{A `data.frame` containing the coordinates and the variable to be used for adding points on the map.}

\item{lab.var}{Variable used for adding points on the map (It should be included in the <dat>).}

\item{lon.var}{Variable name of longitude (It should be included in the <dat>). Default is `longitude`.}

\item{lat.var}{Variable name of latitude (It should be included in the <dat>). Default is `latitude`.}

\item{remove.na}{Whether to remove those rows with `NA` in the <dat>? Default is `FALSE`.}

\item{fill}{A color value or a color vector for filling the points. If the length of <fill> is 1, we directly use such
a color for filling points. If the length of <fill> is greater than 1, then we will render the point by the <lab.var>.
Default is `NA`.}

\item{color}{Color of points. Default is `gray30`.}

\item{alpha}{Transparency of points. Default is `1`.}

\item{hjust}{Adjust of labels at horizontal direction. Default is `0.5`.}

\item{vjust}{Adjust of labels at vertical direction. Default is `0.5`.}
}
\value{
A ggplot2 object.
}
\description{
This function is desinged to add points on the map based on a numeric vector. The shape of point is fixed
as `21`, while the size and color of points could be customized via using the <lab.var> argument for rendering. If you
want to visualize sampling sites on geographic maps, please see the document of `microgeo::add_sampl_site()`.
}
\examples{
##### Example 1: add points on a map with polygons named by grids #####
# Initialization of the visualizations
data(gridded.map.mean)
names(gridded.map.mean)
showtext::showtext_auto(enable = TRUE)

# The size of points represents the average Shannon-Wiener diversity index
gridded.map.mean \%>\% plot_bmap() \%>\%
    add_point(dat = gridded.map.mean@data, lab.var = 'shannon_mean', lon.var = "X.CENTER", lat.var = "Y.CENTER",
              remove.na = TRUE)

# The size of points represents the average Shannon-Wiener diversity index
# There is only one color, so the color has no real significance
gridded.map.mean \%>\% plot_bmap() \%>\%
    add_point(dat = gridded.map.mean@data, lab.var = 'shannon_mean', lon.var = "X.CENTER", lat.var = "Y.CENTER",
              fill = 'lightblue', alpha = 0.5, remove.na = TRUE)

# The size and color of points represents the average Shannon-Wiener diversity index
gridded.map.mean \%>\% plot_bmap() \%>\%
    add_point(dat = gridded.map.mean@data, lab.var = 'shannon_mean', lon.var = "X.CENTER", lat.var = "Y.CENTER",
              fill = c("green", "red"), alpha = 0.5, remove.na = TRUE)
gridded.map.mean \%>\% plot_bmap() \%>\%
    add_point(dat = gridded.map.mean@data, lab.var = 'shannon_mean', lon.var = "X.CENTER", lat.var = "Y.CENTER",
              fill = c("green", "white", "red"), alpha = 0.5, remove.na = TRUE)

# The size and color of points represents the average Shannon-Wiener diversity index
# The number means the standard error of Shannon-Wiener diversity index
gridded.map.mean \%>\% plot_bmap() \%>\%
    add_point(dat = gridded.map.mean@data, lab.var = 'shannon_mean', lon.var = "X.CENTER", lat.var = "Y.CENTER",
              fill = c("green", "white", "red"), alpha = 0.5, remove.na = TRUE) \%>\%
    add_label(dat = gridded.map.mean@data, lab.var = 'shannon_se', lon.var = 'X.CENTER', lat.var = 'Y.CENTER',
              remove.na = TRUE)

# The size and color of points represents the average Shannon-Wiener diversity index
# The label means the standard error of Shannon-Wiener diversity index
gridded.map.mean@data$new.label <- paste0("±", gridded.map.mean@data$shannon_se \%>\% round(., 1))
gridded.map.mean \%>\% plot_bmap() \%>\%
    add_point(dat = gridded.map.mean@data, lab.var = 'shannon_mean', lon.var = "X.CENTER", lat.var = "Y.CENTER",
              fill = c("green", "white", "red"), alpha = 0.5, remove.na = TRUE) \%>\%
    add_label(dat = gridded.map.mean@data, lab.var = 'new.label', lon.var = 'X.CENTER', lat.var = 'Y.CENTER',
              remove.na = TRUE)

# The color of grids means the average Shannon-Wiener diversity index
# The size of points represents the standard error of Shannon-Wiener diversity index
# The number means the number of sampling sites in each grid
gridded.map.mean \%>\% plot_bmap(var = 'shannon_mean', fill = 'auto') \%>\%
    add_point(dat = gridded.map.mean@data, lab.var = 'shannon_se', lon.var = "X.CENTER", lat.var = "Y.CENTER",
              remove.na = TRUE) \%>\%
    add_label(dat = gridded.map.mean@data, lab.var = 'sample.num', lon.var = 'X.CENTER', lat.var = 'Y.CENTER',
              remove.na = FALSE)

##### Example 2: add points on a map with polygons named by administrative district names #####
data(qtp)
met <- cbind(qtp$met, qtp$env)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
map \%>\% plot_bmap() \%>\% add_point(dat = met, lab.var = 'pH')
map \%>\% plot_bmap() \%>\% add_point(dat = met, lab.var = 'pH', fill = 'lightblue', alpha = 0.5)
map \%>\% plot_bmap() \%>\% add_point(dat = met, lab.var = 'pH', fill = RColorBrewer::brewer.pal(11, 'RdYlGn'))
}
\seealso{
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_label]{microgeo::add_label()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
