% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-visualize.R
\name{add_sampl_site}
\alias{add_sampl_site}
\title{Add sampling sites on a map}
\usage{
add_sampl_site(
  map.layer,
  met,
  color.var = NULL,
  color.ord = NULL,
  shape.var = NULL,
  shape.ord = NULL,
  color.val = RColorBrewer::brewer.pal(9, "Set1"),
  fill.val = RColorBrewer::brewer.pal(9, "Set1"),
  shape.val = c(21, 22, 23, 24),
  point.size = 3,
  point.alpha = 0.7,
  facet.by.color = FALSE,
  facet.by.shape = FALSE,
  facet.col.nums = 1,
  facet.title.size = 15
)
}
\arguments{
\item{map.layer}{A ggplot2 object of geographic map.}

\item{met}{A `data.frame` of sample information. Row names must be sample ids, and the column names are observed
variables (e.g., `longitude`, `latitude` and `group`). Longitude and latitude must be included in this dataframe since
we primarily focus on the spatial patterns of microbial traits.}

\item{color.var}{Variable in the <met> used for the color of points. Default is `NULL`.}

\item{color.ord}{Display order of group names rendered by colors. Default is `NULL`.}

\item{shape.var}{Variable in the <met> used for the shape of points. Default is `NULL`.}

\item{shape.ord}{Display order of group names rendered by shapes. Default is `NULL`.}

\item{color.val}{Border (for all shape codes) or fill color (for the shape code less than 21) values of point. Default
is `RColorBrewer::brewer.pal(9, "Set1")`.}

\item{fill.val}{The fill color of point. This argument only works when all shape codes greater than 20. The Default is
`RColorBrewer::brewer.pal(9, "Set1")`.}

\item{shape.val}{Shape values of point. Default is `c(21, 22, 23, 24)`.}

\item{point.size}{Size of point. Default is `3`.}

\item{point.alpha}{Transparency of point. Default is `0.7`.}

\item{facet.by.color}{Whether to facet the graph by color variable? Default is `FALSE`.}

\item{facet.by.shape}{Whether to facet the graph by shape variable? Default is `FALSE`.}

\item{facet.col.nums}{Number of facets. Default is `1`.}

\item{facet.title.size}{Text size of facet title. Default is `15`.}
}
\value{
A ggplot2 object.
}
\description{
This function is designed to add sampling sites on a map by using the \code{ggplot2::geom_sf()} function.
}
\examples{
# Read the map downloaded from DataV.GeoAtlas, and create a standard dataset
data(qtp)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts  <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                               phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%<>\% get_ai(out.dir = "test/microgeo_data")
dataset.dts \%<>\% get_his_bioc(res = 10, out.dir = "test/microgeo_data")
dataset.dts \%<>\% get_soilgrid(out.dir = "test/microgeo_data")
dataset.dts \%<>\% extract_data_from_spatraster(type = 'his')
dataset.dts \%<>\% tidy_dataset()
dataset.dts$met$aridity.region <- cut(dataset.dts$spa$tabs$AI, breaks = c(-Inf, 0.2, 0.5, Inf),
                                      labels = c("H", "A", "S"))
dataset.dts$met$mat.region <- cut(dataset.dts$spa$tabs$Bio1, breaks = c(-Inf, -1, 1, Inf),
                                  labels = c("LT", "MT", "HT"))

# Plot a basic map with sampling sites
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_sampl_site(met = dataset.dts$met, color.val = "gray30", fill.val = "white", point.alpha = 0.5) \%>\%
    add_scale_bar() \%>\% add_north_arrow() \%>\% add_crs()

# Plot a basic map with sampling sites, and add the colors
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_sampl_site(met = dataset.dts$met, color.var = "aridity.region", color.val = "gray30", point.alpha = 0.5) \%>\%
    add_scale_bar() \%>\% add_north_arrow() \%>\% add_crs()

# Plot a basic map with sampling sites, and add the colors/shapes
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_sampl_site(met = dataset.dts$met, color.var = "aridity.region", shape.var = "aridity.region",
                   color.val = "gray30", point.alpha = 0.5) \%>\%
    add_scale_bar() \%>\% add_north_arrow() \%>\% add_crs()

# Plot a basic map, and add the colors/shapes/facet
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_sampl_site(met = dataset.dts$met, color.var = "aridity.region", shape.var = "aridity.region",
                   color.val = "gray30", point.alpha = 0.5, facet.by.color = T, facet.col.nums = 1) \%>\%
    add_scale_bar() \%>\% add_north_arrow() \%>\% add_crs()
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_sampl_site(met = dataset.dts$met, color.var = "aridity.region", shape.var = "mat.region",
                   color.val = "gray30", point.alpha = 0.5, facet.by.color = T, facet.by.shape = T) \%>\%
    add_scale_bar() \%>\% add_north_arrow() \%>\% add_crs()

# Plot a basic map, and add the colors/shapes/facet/SpatRaster
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_spatraster(spat.raster = dataset.dts$spa$rast$his$phh2o) \%>\%
    add_sampl_site(met = dataset.dts$met, color.var = "aridity.region", shape.var = "aridity.region",
                   color.val = "gray30", point.alpha = 0.5, facet.by.color = T, facet.col.nums = 1) \%>\%
    add_scale_bar() \%>\% add_north_arrow() \%>\% add_crs()
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_spatraster(spat.raster = dataset.dts$spa$rast$his$phh2o) \%>\%
    add_sampl_site(met = dataset.dts$met, color.var = "aridity.region", shape.var = "mat.region",
                   color.val = "gray30", point.alpha = 0.5, facet.by.color = T, facet.by.shape = T) \%>\%
    add_scale_bar() \%>\% add_north_arrow() \%>\% add_crs()
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:get_ai]{microgeo::get_ai()}}
\code{\link[microgeo:get_his_bioc]{microgeo::get_his_bioc()}}
\code{\link[microgeo:get_soilgrid]{microgeo::get_soilgrid()}}
\code{\link[microgeo:extract_data_from_spatraster]{microgeo::extract_data_from_spatraster()}}
\code{\link[microgeo:tidy_dataset]{microgeo::tidy_dataset()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[microgeo:add_crs]{microgeo::add_crs()}}
\code{\link[microgeo:add_spatraster]{microgeo::add_spatraster()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
