% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-visualize.R
\name{add_scale_bar}
\alias{add_scale_bar}
\title{Add spatial-aware scale bar on a map}
\usage{
add_scale_bar(
  map.layer,
  position = c("bl", "br", "tl", "tr"),
  fill = c("white", "black"),
  size = 1,
  color = "black",
  height = 0.4,
  hint = 0.5,
  ...
)
}
\arguments{
\item{map.layer}{A ggplot2 object of geographic map.}

\item{position}{Position of spatial-aware scale bar. Select one from `'bl', 'br', 'tl', 'tr'`. Default is `bl`.}

\item{fill}{Fill color of spatial-aware scale bar. Two color values are required. Default is `c('white', 'black')`.}

\item{size}{Size of text in spatial-aware scale bar. Default is `1`.}

\item{color}{Color of text in spatial-aware scale bar. Default is `black`.}

\item{height}{Height of spatial-aware scale bar. Default is `0.4` cm.}

\item{hint}{Proportion of the plot area which the spatial-aware scale bar should occupy. Default is `0.4`.}

\item{...}{Parameters parsed by \code{ggspatial::annotation_scale()}.}
}
\value{
A ggplot2 object.
}
\description{
This function is used to add spatial-aware scale bar on a map using \code{ggspatial::annotation_scale()}.
}
\examples{
data(common.map.mean)
common.map.mean  \%>\% plot_bmap()
common.map.mean  \%>\% plot_bmap() \%>\% add_north_arrow()
common.map.mean  \%>\% plot_bmap() \%>\% add_north_arrow() \%>\% add_scale_bar()
}
\seealso{
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[ggspatial:annotation_scale]{ggspatial::annotation_scale()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
