% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-visualize.R
\name{add_spatraster}
\alias{add_spatraster}
\title{Add SpatRaster on a map}
\usage{
add_spatraster(
  map.layer,
  spat.raster,
  colors = "auto",
  breaks = NULL,
  labels = NULL,
  show.border = TRUE,
  border.color = "black",
  border.size = 1,
  legend.title = NULL,
  facet.col.nums = 1,
  facet.title.size = 15,
  ...
)
}
\arguments{
\item{map.layer}{A ggplot2 object of geographic map.}

\item{spat.raster}{A `SpatRaster` object to be visualized.}

\item{colors}{A vector of color values, which would be used for SpatRaster visualization. Default is `'auto'`.}

\item{breaks}{A numeric vector indicating the breaks of legend. If it is `NULL`, then a continuous color bar of legend
would be automatically applied. Default is `NULL`.}

\item{labels}{A character vector indicating the labels of legend break. Only works when breaks is not `NULL`.}

\item{show.border}{Whether to show map border on the SpatRaster? Default is `TRUE`.}

\item{border.color}{Color of map border. Default is `black`.}

\item{border.size}{Width of map border. Default is `1`.}

\item{legend.title}{The title of legend. Default is `NULL`.}

\item{facet.col.nums}{Number of facet. Only works when there are more than one layers in SpatRaster. Default is `1`.}

\item{facet.title.size}{Text size of facet title. It only works when there are more than one layers in SpatRaster. The
Default is `15`.}

\item{...}{Parameters parsed by \code{tidyterra::geom_spatraster()}.}
}
\value{
A ggplot2 object.
}
\description{
This function is designed to add a SpatRaster on the map via using the \code{tidyterra::geom_spatraster}.
}
\examples{
# Read the map downloaded from DataV.GeoAtlas, and create a standard dataset
data(qtp)
showtext::showtext_auto(enable = TRUE)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts  <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                               phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%<>\% get_his_bioc(res = 2.5, out.dir = "test/microgeo_data")
dataset.dts \%<>\% get_fut_bioc(res = 2.5, out.dir = "test/microgeo_data")
dataset.dts \%<>\% get_soilgrid(depth = 5, out.dir = "test/microgeo_data", measures = c('phh2o'))
dataset.dts \%>\% show_dataset()

# Add SpatRaster on a map without facets (`soil pH` without legend breaks)
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_spatraster(spat.raster = dataset.dts$spa$rast$his$phh2o) \%>\%
    add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER')

# Add SpatRaster on a map without facets (`soil pH` with custom legend breaks)
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_spatraster(spat.raster = dataset.dts$spa$rast$his$phh2o, breaks = c(5.5, 6.5, 7.5, 8.5)) \%>\%
    add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER')

# Add SpatRaster on a map without facets (`soil pH` with custom legend breaks and labels)
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_spatraster(spat.raster = dataset.dts$spa$rast$his$phh2o, breaks = c(5,6,7,8),
                   labels = c("A", "B", "C", "D", "E")) \%>\%
    add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER')

# Add SpatRaster on a map with two facets (`Mean annual temperature` without legend breaks)
spatDat <- c(dataset.dts$spa$rast$his$Bio1, dataset.dts$spa$rast$fut$`BCC-CSM2-MR|ssp585|2061-2080`$Bio1)
names(spatDat) <- c("Historical", "ssp585|2061-2080")
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_spatraster(spat.raster = spatDat, legend.title = "Bio1", facet.col.nums = 2) \%>\%
    add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER')

# Add SpatRaster on a map with two facets (`Mean annual temperature` with custom legend breaks and labels)
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_spatraster(spat.raster = spatDat, breaks = c(-2,-1,1,2), labels = c("A", "B", "C", "D", "E"),
                   legend.title = "Bio1", facet.col.nums = 2) \%>\%
    add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER')
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:get_his_bioc]{microgeo::get_his_bioc()}}
\code{\link[microgeo:get_fut_bioc]{microgeo::get_fut_bioc()}}
\code{\link[microgeo:get_soilgrid]{microgeo::get_soilgrid()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_label]{microgeo::add_label()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
