% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-biogeo-trait.R
\name{calc_markers}
\alias{calc_markers}
\title{Identify ecological markers highly correlated with one or mutiple abiotic properties}
\usage{
calc_markers(
  dataset,
  use.var,
  use.dat = c("env", "spa"),
  annotation.level = "Phylum",
  cor.method = c("pearson", "spearman", "kendall"),
  dist.method = "bray",
  p.adjust.method = "none",
  r.thres = 0.1,
  p.thres = 0.05,
  mantel.permutations = 999,
  mantel.parallel = 20
)
}
\arguments{
\item{dataset}{A microgeo dataset with the class of `MicrogeoDataset`.}

\item{use.var}{Variables used for correlations. If there are only one variable, then the correlation coefficient would
be calculated by using \code{psych::corr.test()}. If there are mutiple variables, the correlation coefficient would be
calculated by using \code{vegan::mantel()}.}

\item{use.dat}{Which type of data would be used for correlations? Select one from `env` (measured properties) and `spa`
(the data extracted from spatial dataset). Default is `env`.}

\item{annotation.level}{Which annotation level would be used? Default is `Phylum`.}

\item{cor.method}{Correlation method, as accepted by cor: `pearson`, `spearman` or `kendall`. Default is `pearson`.}

\item{dist.method}{Method for distance calculations if there are mutiple variables (<use.var>). Default is `bray`. See
\code{vegan::vegdist()} for more options.}

\item{p.adjust.method}{A method to adjust P value. Select from `holm`, `hochberg`, `hommel`, `bonferroni`, `BH`, `BY`,
`fdr`, `none`. Default is `none`.}

\item{r.thres}{Threshold of correlation coefficient to identify markers. Default is `0.1`.}

\item{p.thres}{Threshold of P value to identify markers. Default is `0.05`.}

\item{mantel.permutations}{Permutations for mantel test if there are mutiple variables (<use.var>). Default is `999`.}

\item{mantel.parallel}{Parallel number for mantel test if there are mutiple variables (<use.var>). Default is `20`.}
}
\value{
A `MicrogeoDataset` class with the following components:
\describe{
  \item{\code{object$mat}}{A `data.frame` of ASV/gene abundance.}
  \item{\code{object$ant}}{A `data.frame` of ASV/gene anootations.}
  \item{\code{object$met}}{A `data.frame` of sample information.}
  \item{\code{object$map}}{A `SpatialPolygonsDataFrame` of map.}
  \item{\code{object$phy}}{A phylogenetic tree with `newick` format if applicable.}
  \item{\code{object$env}}{A `data.frame` of measured environmental properties if applicable.}
  \item{\code{object$abd$raw}}{A `list` of taxonomic/gene abundance tables.}
  \item{\code{object$abd$mar}}{A `list` of taxonomic/gene marker tables.}
  \item{\code{object$*}}{Spatial and biogeographic traits if applicable.}
}
}
\description{
This function is implemented to identify ecological markers highly correlate with one or mutiple abiotic
properties. Particularly, it requires the relative abundance data calculated by \code{microgeo::calc_rel_abund()}.
}
\examples{
# Create a microgeo dataset
data(qtp)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()

# Extract 19 historical bioclimatic variables
dataset.dts \%<>\% get_his_bioc(res = 2.5, out.dir = "test/microgeo_data")
dataset.dts \%<>\% extract_data_from_spatraster(type = 'his')
dataset.dts \%>\% show_dataset()

# Tidy up microgeo dataset
dataset.dts \%<>\% rarefy_count_table()
dataset.dts \%<>\% tidy_dataset()
dataset.dts \%>\% show_dataset()

# Calculate relative abundance
dataset.dts \%<>\% calc_rel_abund()
dataset.dts \%>\% show_dataset()

# Identify ecological markers based on soil pH in <env>
# Correlation coefficients would be calculated by `psych::corr.test()`
dataset.dts \%<>\% calc_markers(use.var = 'pH', annotation.level = 'Phylum', r.thres = 0.1)
dataset.dts \%>\% show_dataset()
head(dataset.dts$abd$mar$correlation)

# Identify ecological markers based on Bio12 in <spa>
# Correlation coefficients would be calculated by `psych::corr.test()`
dataset.dts \%<>\% calc_markers(use.var = 'Bio12', use.dat = 'spa', annotation.level = 'Phylum', r.thres = 0.1)
dataset.dts \%>\% show_dataset()
head(dataset.dts$abd$mar$correlation)

# Identify ecological markers based on soil pH and TOC in <env>
# Correlation coefficients would be calculated by `vegan::mantel()`
dataset.dts \%<>\% calc_markers(use.var = c('pH', 'TOC'), annotation.level = 'Phylum', r.thres = 0.1)
dataset.dts \%>\% show_dataset()
head(dataset.dts$abd$mar$correlation)
}
\seealso{
\code{\link[psych:corr.test]{psych::corr.test()}}
\code{\link[vegan:mantel]{vegan::mantel()}}
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:get_his_bioc]{microgeo::get_his_bioc()}}
\code{\link[microgeo:extract_data_from_spatraster]{microgeo::extract_data_from_spatraster()}}
\code{\link[microgeo:rarefy_count_table]{microgeo::rarefy_count_table()}}
\code{\link[microgeo:tidy_dataset]{microgeo::tidy_dataset()}}
\code{\link[microgeo:calc_rel_abund]{microgeo::calc_rel_abund()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
