% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-biogeo-trait.R
\name{calc_rel_abund}
\alias{calc_rel_abund}
\title{Calculate relatively taxonomic/gene abundance}
\usage{
calc_rel_abund(dataset, use.perct = TRUE)
}
\arguments{
\item{dataset}{A microgeo dataset with the class of `MicrogeoDataset`.}

\item{use.perct}{Whether to use the percentages for relatively taxonomic/gene abundance. Default is `TRUE`.}
}
\value{
A `MicrogeoDataset` class with the following components:
\describe{
  \item{\code{object$mat}}{A `data.frame` of ASV/gene abundance.}
  \item{\code{object$ant}}{A `data.frame` of ASV/gene anootations.}
  \item{\code{object$met}}{A `data.frame` of sample information.}
  \item{\code{object$map}}{A `SpatialPolygonsDataFrame` of map.}
  \item{\code{object$phy}}{A phylogenetic tree with `newick` format if applicable.}
  \item{\code{object$env}}{A `data.frame` of measured environmental properties if applicable.}
  \item{\code{object$abd$raw}}{A `list` of taxonomic/gene abundance tables.}
  \item{\code{object$*}}{Spatial and biogeographic traits if applicable.}
}
}
\description{
This function is used to calculate the relatively taxonomic/gene abundance based on classification levels.
}
\examples{
# Create a microgeo dataset
data(qtp)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()

# Tidy up microgeo dataset
dataset.dts \%<>\% rarefy_count_table()
dataset.dts \%<>\% tidy_dataset()
dataset.dts \%>\% show_dataset()

# Calculate relative abundance
dataset.dts \%<>\% calc_rel_abund()
dataset.dts \%>\% show_dataset()

# Check the results
head(dataset.dts$abd$raw$Phylum[,1:5])
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:rarefy_count_table]{microgeo::rarefy_count_table()}}
\code{\link[microgeo:tidy_dataset]{microgeo::tidy_dataset()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
