% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microgeo-dataset.R
\name{check_map_area}
\alias{check_map_area}
\title{Check the map area}
\usage{
check_map_area(map, met, lon = "longitude", lat = "latitude")
}
\arguments{
\item{map}{A microgeo-compatible `SpatialPolygonsDataFrame` of map. See \code{microgeo::trans_map_fmt()} for details.}

\item{met}{A `data.frame` of sample information. Row names should be sample ids and the column names must be variables
(e.g., `longitude`, `latitude` and `group`). Longitude and latitude must be included in this `data.frame` as we mainly
focus on the spatial patterns of microbial traits.}

\item{lon}{Column name of longitude in `met`. Default is `longitude`.}

\item{lat}{Column name of latitude in `met`. Default is `latitude`.}
}
\value{
A `data.frame` of the coordinates for sampling sites exceeded the map area.
}
\description{
This function is designed to check whether all sampling sites fall within the map area.
}
\examples{
##### Example 1: a map downloaded from the DataV.GeoAtlas ####
data(qtp)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
rst <- map \%>\% check_map_area(met = qtp$met, lon = 'longitude', lat = 'latitude')
nrow(rst) # it is expected to be 0

##### Example 2: a map loaded from a local ESRI Shapefile (QTP) ####
map <- system.file("shapefiles/qtp-map", "DBATP_Polygon.shp", package = "microgeo") \%>\% read_shp_map()
map@data$NAME <- "Qinghai-Tibet Plateau" # Add a column to meet the demands of microgeo::trans_map_fmt()
map \%<>\% trans_map_fmt(var = 'NAME')
rst <- map \%>\% check_map_area(met = qtp$met, lon = 'longitude', lat = 'latitude')
nrow(rst) # it is expected to be 3
map \%>\% plot_bmap() \%>\% add_sampl_site(met = rst) \%>\%
    add_scale_bar() \%>\% add_north_arrow() \%>\% add_crs() # Plot a map to show the location of these samples on map
# Three samples locate on the boundary of map
# Thus, we can ignore the warnings during creating microgeo dataset, if we use the current map.
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:read_shp_map]{microgeo::read_shp_map()}}
\code{\link[microgeo:trans_map_fmt]{microgeo::trans_map_fmt()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_sampl_site]{microgeo::add_sampl_site()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[microgeo:add_crs]{microgeo::add_crs()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
