% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microgeo-dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{Create a standard microgeo dataset for biogeographic analysis}
\usage{
create_dataset(
  mat,
  ant,
  met,
  map,
  phy = NULL,
  env = NULL,
  lon = "longitude",
  lat = "latitude"
)
}
\arguments{
\item{mat}{A `data.frame` of ASV/gene abundances. Row names must be ASV/gene ids and the column names should be sample
ids. To avoid unexpected errors, sample ids and ASV/gene ids should not be start with a number. The same below.}

\item{ant}{A `data.frame` of ASV/gene annotations. Rows names must be ASV/gene ids and the column names are annotation
levels (e.g., `Kingdom`, `Phylum`, `Class`, `Order`, `Family`, `Genus`, `Species` for ASV table; `Level 1`, `Level 2`,
`Level 3`, `Level 4` for gene abundance table).}

\item{met}{A `data.frame` of sample information. Row names should be sample ids and the column names must be variables
(e.g., `longitude`, `latitude` and `group`). Longitude and latitude must be included in this `data.frame` as we mainly
focus on the spatial patterns of microbial traits.}

\item{map}{A microgeo-compatible `SpatialPolygonsDataFrame` of map. See \code{microgeo::trans_map_fmt()} for details.}

\item{phy}{An optional phylogenetic tree with a `newick` format (`phylo` class). Required for calculating phylogenetic
traits. Default is `NULL`.}

\item{env}{An optional `data.frame` of measured environmental properties (e.g., `pH` and `TOC`). The row names must be
sample ids and the column names should be measured variables. Default is `NULL`.}

\item{lon}{Column name of longitude in `met`. Default is `longitude`.}

\item{lat}{Column name of latitude in `met`. Default is `latitude`.}
}
\value{
A `MicrogeoDataset` class with the following components:
\describe{
  \item{\code{object$mat}}{A `data.frame` of ASV/gene abundance.}
  \item{\code{object$ant}}{A `data.frame` of ASV/gene anootations.}
  \item{\code{object$met}}{A `data.frame` of sample information.}
  \item{\code{object$map}}{A `SpatialPolygonsDataFrame` of map.}
  \item{\code{object$phy}}{A phylogenetic tree with `newick` format (`phylo` class) if applicable.}
  \item{\code{object$env}}{A `data.frame` of measured environmental properties if applicable.}
}
}
\description{
This function wraps the methods for dataset preparation. It requires several microbiome dataset including
ASV (also referred to as OTU in QIIME 1) or gene abundance table (e.g., gene abundance tables generated by metagenomic
sequencing, gene chip or high-throughput qPCR techniques), taxonomic/functional annotation tables, metadata containing
the coordinates of sampling site (longitude & latitude; required) and other information such as the grouping schema of
sampling site, phylogenetic tree of ASVs, and measured environmental properties like `pH` and `TOC`. In addition, this
function also requires a `SpatialPolygonsDataFrame` of map because we mainly focus on the spatial pattern of microbial
traits. Phylogenetic tree and measured environmental properties are optional.
}
\examples{
# Read a map from the DataV.GeoAtlas
showtext::showtext_auto(enable = TRUE)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
map \%>\% plot_bmap() \%>\% add_label(dat = map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
    add_scale_bar() \%>\% add_north_arrow() \%>\% add_crs()

# Create a microgeo dataset using bacterial data and the map retrieved from the DataV.GeoAtlas
data(qtp)
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_label]{microgeo::add_label()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[microgeo:add_crs]{microgeo::add_crs()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
