% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-ml-model.R
\name{evaluate_ml_model}
\alias{evaluate_ml_model}
\title{Evaluate machine learning model created by \code{microgeo::create_ml_model()}}
\usage{
evaluate_ml_model(model.dat, only.show.class = TRUE)
}
\arguments{
\item{model.dat}{A machine learning model object returned by \code{microgeo::create_ml_model()}.}

\item{only.show.class}{Only show observed classifications in ROC curves (delete micro-/macro- average)? Only works for
classification model. Default is `TRUE`.}
}
\value{
A `ggplot2` object.
}
\description{
This function is desinged to evaluate a ML model created by \code{microgeo::create_ml_model()}.
}
\examples{
# Create a microgeo dataset
data(qtp)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()

# Download aridity index and 19 historical bioclimatic variables
dataset.dts \%<>\% get_ai(out.dir = "test/microgeo_data")
dataset.dts \%<>\% get_his_bioc(res = 2.5, out.dir = "test/microgeo_data")
dataset.dts \%<>\% extract_data_from_spatraster(type = 'his')

# Tidy up microgeo dataset
dataset.dts \%<>\% rarefy_count_table()
dataset.dts \%<>\% tidy_dataset()
dataset.dts \%>\% show_dataset()

# Calculate the relative abundance and the ecologycal markers at `Family` level
dataset.dts \%<>\% calc_rel_abund()
dataset.dts \%<>\% calc_markers(use.dat = 'spa', use.var = 'AI', annotation.level = 'Family', r.thres = 0.3)
head(dataset.dts$abd$mar$correlation)

# Create a regression random forest model for soil pH
rf.rst.reg <- create_ml_model(y.data = dataset.dts$env,
                              x.data = dataset.dts$spa$tabs[,paste0("Bio", seq(19))], var = 'pH', method = 'rf')
rf.rst.reg \%>\% evaluate_ml_model()

# Create a binary classification random forest model for the family of f__A21b
family.bins <- data.frame(row.names = rownames(dataset.dts$abd$mar$abundance),
                          f__A21b = dataset.dts$abd$mar$abundance$f__A21b)
family.bins$f__A21b <- ifelse(family.bins$f__A21b > 0, "presence", "absence") # two classifications
family.bins$f__A21b <- as.factor(family.bins$f__A21b)
rf.rst.cla.bin <- create_ml_model(y.data = family.bins,
                                  x.data = dataset.dts$spa$tabs[,paste0("Bio", seq(19))],
                                  var = 'f__A21b', method = 'rf', type = 'classification')
rf.rst.cla.bin \%>\% evaluate_ml_model()

# Create a mutiple class classification random forest model for the family of f__A21b
family.mutiple <- data.frame(row.names = rownames(dataset.dts$abd$mar$abundance),
                             f__A21b = dataset.dts$abd$mar$abundance$f__A21b)
family.mutiple$f__A21b <- cut(family.mutiple$f__A21b, breaks = c(-Inf, 0.05, 0.2, 0.9, Inf),
                              labels = c("H", "A", "S", "Y")) # four classifications
family.mutiple$f__A21b <- as.factor(family.mutiple$f__A21b)
rf.rst.cla.mutiple <- create_ml_model(y.data = family.mutiple,
                                      x.data = dataset.dts$spa$tabs[,paste0("Bio", seq(19))],
                                      var = 'f__A21b', method = 'rf', type = 'classification')
rf.rst.cla.mutiple \%>\% evaluate_ml_model()
}
\seealso{
\code{\link[psych:corr.test]{psych::corr.test()}}
\code{\link[multiROC:multi_roc]{multiROC::multi_roc()}}
\code{\link[multiROC:plot_roc_data]{multiROC::plot_roc_data()}}
\code{\link[microgeo:create_ml_model]{microgeo::create_ml_model()}}
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:get_ai]{microgeo::get_ai()}}
\code{\link[microgeo:get_his_bioc]{microgeo::get_his_bioc()}}
\code{\link[microgeo:extract_data_from_spatraster]{microgeo::extract_data_from_spatraster()}}
\code{\link[microgeo:rarefy_count_table]{microgeo::rarefy_count_table()}}
\code{\link[microgeo:tidy_dataset]{microgeo::tidy_dataset()}}
\code{\link[microgeo:calc_rel_abund]{microgeo::calc_rel_abund()}}
\code{\link[microgeo:calc_markers]{microgeo::calc_markers()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
