% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-operate.R
\name{extract_metadata_from_map}
\alias{extract_metadata_from_map}
\title{Extract the metadata table from a geographic map}
\usage{
extract_metadata_from_map(map, met)
}
\arguments{
\item{map}{A geographic map with the class of `SpatialPolygonsDataFrame`.}

\item{met}{Original metadata table. See the examples for more details about such a table.}
}
\value{
A `data.frame` of new metadata table extracted from the geographic map.
}
\description{
This function is designed to extract metadata table from a geographic map. The results returned by such a
function would be useful for the statistical analysis based on administrative regions or grid names.
}
\examples{
##### Example 1: extract the metadata from a map retrieved from the DataV.GeoAtlas #####
data(qtp)
names(qtp)
head(qtp$met)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
metadata <- map \%>\% extract_metadata_from_map(met = qtp$met)
head(metadata) # Now we can compare any metrics among administrative regions based on such a new metadata table

##### Example 2: extract the metadata from a gridded map #####
map \%<>\% grid_map(res = 1.5)
metadata <- map \%>\% extract_metadata_from_map(met = qtp$met)
head(metadata) # Now we can compare any metrics among grids based on such a new metadata table

##### Example 3: extract the metadata from a map with additional data #####
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%<>\% rarefy_count_table()
dataset.dts \%<>\% tidy_dataset()
dataset.dts \%<>\% calc_alpha_div(measures = c("observed", "shannon"))
dataset.dts \%>\%  show_dataset()
map <- merge_dfs_to_map(map = dataset.dts$map, dat = dataset.dts$div$alpha,
                        met = dataset.dts$met, agg.method = 'mean')
metadata <- map \%>\% extract_metadata_from_map(met = dataset.dts$met)
head(metadata) # Now we can compare the alpha diversity metrics among administrative regions
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:grid_map]{microgeo::grid_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:rarefy_count_table]{microgeo::rarefy_count_table()}}
\code{\link[microgeo:tidy_dataset]{microgeo::tidy_dataset()}}
\code{\link[microgeo:calc_alpha_div]{microgeo::calc_alpha_div()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:merge_dfs_to_map]{microgeo::merge_dfs_to_map()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
