% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-spatial-trait.R
\name{get_fut_bioc}
\alias{get_fut_bioc}
\title{Download future bioclimatic variables from WorldClim database}
\usage{
get_fut_bioc(
  dataset,
  gcm = c("BCC-CSM2-MR", "ACCESS-CM2", "CNRM-CM6-1", "CNRM-ESM2-1", "CanESM5"),
  sce = c("ssp126", "ssp585"),
  yea = c("2021-2040", "2061-2080"),
  res = c(10, 5, 2.5),
  out.dir = "microgeo_data"
)
}
\arguments{
\item{dataset}{A microgeo dataset with the class of `MicrogeoDataset`.}

\item{gcm}{Which model abbreviation of future bioclimatic data would be used? Default is `BCC-CSM2-MR`. There are many
climate model are avaliable, Visit `https://www.worldclim.org/data/cmip6/cmip6climate.html` for more details.}

\item{sce}{Which Shared Socio-economic Pathway code of future bioclimatic data would be applied? Select one or mutiple
from `ssp126, ssp245, ssp370, ssp585`.}

\item{yea}{Which time period would be applied? Select from `2021-2040, 2041-2060, 2061-2080, 2081-2100`.}

\item{res}{Which resolution would be used? Valid values are `10, 5, 2.5, 0.5` (minutes of a degree). Default is `10`.}

\item{out.dir}{Directory path to save downloaded files. Default is `microgeo_data`}
}
\value{
A `MicrogeoDataset` class with the following components:
\describe{
  \item{\code{object$mat}}{A `data.frame` of ASV/gene abundance.}
  \item{\code{object$ant}}{A `data.frame` of ASV/gene anootations.}
  \item{\code{object$met}}{A `data.frame` of sample information.}
  \item{\code{object$map}}{A `SpatialPolygonsDataFrame` of map.}
  \item{\code{object$phy}}{A phylogenetic tree with `newick` format if applicable.}
  \item{\code{object$env}}{A `data.frame` of measured environmental properties if applicable.}
  \item{\code{object$spa$rast$fut}}{Future `SpatRaster` of bioclimatic variables.}
  \item{\code{object$*}}{Spatial and biogeographic traits if applicable.}
}
}
\description{
This function is implemented to download future bioclimatic variables from WorldClim database version 2.1
(`https://www.worldclim.org/`).
}
\examples{
# Create a microgeo dataset
data(qtp)
showtext::showtext_auto(enable = TRUE)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()

# Download 19 historical and future bioclimatic variables for the research region
dataset.dts \%<>\% get_his_bioc(res = 2.5, out.dir = "test/microgeo_data")
dataset.dts \%<>\% get_fut_bioc(res = 2.5, out.dir = "test/microgeo_data")
dataset.dts \%>\% show_dataset()

# Visualize the Bio12 (Mean annual precipitation) of `BCC-CSM2-MR|ssp585|2061-2080`. Just an example.
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$fut$`BCC-CSM2-MR|ssp585|2061-2080`$Bio12,
                  color = colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100)) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$fut$`BCC-CSM2-MR|ssp585|2061-2080`$Bio12,
                  breaks = c(100, 300, 500, 700),
                  color = RColorBrewer::brewer.pal(11, "RdYlGn")[c(1,3,5,9,11)],
                  labels = c("A", "B", "C", "D", "E")) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()

# Visualize the Bio1 (Mean annual temperature) of `BCC-CSM2-MR|ssp585|2061-2080`. Just an example.
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$fut$`BCC-CSM2-MR|ssp585|2061-2080`$Bio1,
                  color = colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100)) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$fut$`BCC-CSM2-MR|ssp585|2061-2080`$Bio1,
                  breaks = c(-2, -1, 1, 2),
                  color = RColorBrewer::brewer.pal(11, "RdYlGn")[c(1,3,5,9, 11)],
                  labels = c("A", "B", "C", "D", "E")) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:get_his_bioc]{microgeo::get_his_bioc()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_spatraster]{microgeo::add_spatraster()}}
\code{\link[microgeo:add_label]{microgeo::add_label()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
\code{\link[microgeo:add_crs]{microgeo::add_crs()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
