% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-spatial-trait.R
\name{get_his_bioc}
\alias{get_his_bioc}
\title{Download historical bioclimatic variables from WorldClim database}
\usage{
get_his_bioc(dataset, res = c(10, 5, 2.5, 0.5), out.dir = "microgeo_data")
}
\arguments{
\item{dataset}{A microgeo dataset with the class of `MicrogeoDataset`.}

\item{res}{Which resolution would be used? Valid values are `10, 5, 2.5, 0.5` (minutes of a degree). Default is `10`.}

\item{out.dir}{Directory path to save downloaded files. Default is `microgeo_data`}
}
\value{
A `MicrogeoDataset` class with the following components:
\describe{
  \item{\code{object$mat}}{A `data.frame` of ASV/gene abundance.}
  \item{\code{object$ant}}{A `data.frame` of ASV/gene anootations.}
  \item{\code{object$met}}{A `data.frame` of sample information.}
  \item{\code{object$map}}{A `SpatialPolygonsDataFrame` of map.}
  \item{\code{object$phy}}{A phylogenetic tree with `newick` format if applicable.}
  \item{\code{object$env}}{A `data.frame` of measured environmental properties if applicable.}
  \item{\code{object$spa$rast$his}}{Historical `SpatRaster` of bioclimatic variables and other vars if applicable.}
  \item{\code{object$*}}{Spatial and biogeographic traits if applicable.}
}
}
\description{
This function is implemented to download historical bioclimatic variables from WorldClim database version
2.1 (`https://www.worldclim.org/`).
}
\examples{
# Create a microgeo dataset
data(qtp)
showtext::showtext_auto(enable = TRUE)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()

# Download 19 historical bioclimatic variables for the research region
dataset.dts \%<>\% get_his_bioc(res = 10, out.dir = "test/microgeo_data")
dataset.dts \%>\% show_dataset()

# Visualize the Bio12 (Mean annual precipitation). Just an example.
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$his$Bio12,
                  color = colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100)) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$his$Bio12, breaks = c(100, 300, 500, 700),
                  color = RColorBrewer::brewer.pal(11, "RdYlGn")[c(1,3,5,9,11)],
                  labels = c("A", "B", "C", "D", "E")) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()

# Visualize the Bio1 (Mean annual temperature). Just an example.
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$his$Bio1,
                  color = colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100)) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$his$Bio1, breaks = c(-2, -1, 1, 2),
                  color = RColorBrewer::brewer.pal(11, "RdYlGn")[c(1,3,5,9,11)],
                  labels = c("A", "B", "C", "D", "E")) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_spatraster]{microgeo::add_spatraster()}}
\code{\link[microgeo:add_label]{microgeo::add_label()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
\code{\link[microgeo:add_crs]{microgeo::add_crs()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
