% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-spatial-trait.R
\name{get_modis_cla_metrics}
\alias{get_modis_cla_metrics}
\title{Download classification MODIS metrics from EOSDIS}
\usage{
get_modis_cla_metrics(
  dataset,
  username,
  password,
  measures = "LC_Type1",
  year = "least",
  nums.job = 2,
  out.dir = "microgeo_data"
)
}
\arguments{
\item{dataset}{A microgeo dataset with the class of `MicrogeoDataset`.}

\item{username}{Username of your (free) EOSDIS account. Sign up \href{https://urs.earthdata.nasa.gov/users/new}{here}.}

\item{password}{Password of your (free) EOSDIS account. Sign up \href{https://urs.earthdata.nasa.gov/users/new}{here}.}

\item{measures}{Measures. Use \code{\link[microgeo:show_modis_cla_metrics]{microgeo::show_modis_cla_metrics()}} to see
all avaliable  measures. More details are avaliable at \href{https://lpdaac.usgs.gov/products/mcd12q1v006/}{MODIS} and
\href{https://lpdaac.usgs.gov/documents/1409/MCD12_User_Guide_V61.pdf}{MCD12_User_Guide_V61}. Default is `"LC_Type1"`.}

\item{year}{Download data for which year? If it is `least`, we will search data by using value obtained by subtracting
two years from the current year. Default is `least`.}

\item{nums.job}{How many threads would be used for the remote-sense image merging? Default is `2`. This would consume
a large amounts of computational resources!}

\item{out.dir}{Directory path to save downloaded files. Default is `microgeo_data`}
}
\value{
A `MicrogeoDataset` class with the following components:
\describe{
  \item{\code{object$mat}}{A `data.frame` of ASV/gene abundance.}
  \item{\code{object$ant}}{A `data.frame` of ASV/gene anootations.}
  \item{\code{object$met}}{A `data.frame` of sample information.}
  \item{\code{object$map}}{A `SpatialPolygonsDataFrame` of map.}
  \item{\code{object$phy}}{A phylogenetic tree with `newick` format if applicable.}
  \item{\code{object$env}}{A `data.frame` of measured environmental properties if applicable.}
  \item{\code{object$spa$rast$cla}}{Historical `SpatRaster` of classification MODIS metrics.}
  \item{\code{object$*}}{Spatial and biogeographic traits if applicable.}
}
}
\description{
This function is implemented to download some classification MODIS metrics from the EOSDIS website ( URL:
`https://cmr.earthdata.nasa.gov` ). When visualizing the `SpatRaster` obtained from this function, we do not recommend
using `add_crs()`. This is because the data has a high resolution, which would significantly slow down the plotting.
}
\examples{
# Create a microgeo dataset
data(qtp)
showtext::showtext_auto(enable = TRUE)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()

# Download classification MODIS metrics of research region
dataset.dts \%<>\% get_modis_cla_metrics(username = "bioape.lichaonan",
                                       password = "Tipo@cib.0905", out.dir = "test/microgeo_data")
dataset.dts \%>\% show_dataset()

# Visualize LC_Type1
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$cla$LC_Type1,
                  color = c(RColorBrewer::brewer.pal(12, "Set3"), RColorBrewer::brewer.pal(9, "Set1"))) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar()
}
\seealso{
\code{\link[microgeo:show_modis_cla_metrics]{microgeo::show_modis_cla_metrics()}}
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_spatraster]{microgeo::add_spatraster()}}
\code{\link[microgeo:add_label]{microgeo::add_label()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
