% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-spatial-trait.R
\name{get_modis_num_metrics}
\alias{get_modis_num_metrics}
\title{Download numeric MODIS metrics from EOSDIS}
\usage{
get_modis_num_metrics(
  dataset,
  username,
  password,
  measures = c("NDVI", "EVI"),
  prod.res = c(1000, 500, 250),
  prod.typ = c("Terra", "Aqua"),
  date.ran = c("2019-08-01|2019-09-01", "2020-08-01|2020-09-01"),
  nums.job = 30,
  out.dir = "microgeo_data"
)
}
\arguments{
\item{dataset}{A microgeo dataset with the class of `MicrogeoDataset`.}

\item{username}{Username of your (free) EOSDIS account. Sign up \href{https://urs.earthdata.nasa.gov/users/new}{here}.}

\item{password}{Password of your (free) EOSDIS account. Sign up \href{https://urs.earthdata.nasa.gov/users/new}{here}.}

\item{measures}{Measures. Use \code{\link[microgeo:show_modis_num_metrics]{microgeo::show_modis_num_metrics()}} to see
all avaliable measures. More details are avaliable at \href{https://modis.gsfc.nasa.gov/data/}{MODIS Web}. The default
is all avaliable measures.}

\item{prod.res}{Spatial resolution of MODIS image product. Valid resolution are `1000, 500, 250`(meter). It only works
when the measures are `NDVI` and/or `ENV`. Other MODIS image product harbor a fixed resolution. Default is `1000` m.}

\item{prod.typ}{Type of MODIS image product. Select one from `"Terra", "Aqua"`. Default is `Terra`.}

\item{date.ran}{Which date range would be applied? Default is `c("2019-08-01|2019-09-01", "2020-08-01|2020-09-01")`.}

\item{nums.job}{How many threads would be used for the remote-sense image merging? Default is `30`. This would consume
a large amounts of computational resources!}

\item{out.dir}{Directory path to save downloaded files. Default is `microgeo_data`}
}
\value{
A `MicrogeoDataset` class with the following components:
\describe{
  \item{\code{object$mat}}{A `data.frame` of ASV/gene abundance.}
  \item{\code{object$ant}}{A `data.frame` of ASV/gene anootations.}
  \item{\code{object$met}}{A `data.frame` of sample information.}
  \item{\code{object$map}}{A `SpatialPolygonsDataFrame` of map.}
  \item{\code{object$phy}}{A phylogenetic tree with `newick` format if applicable.}
  \item{\code{object$env}}{A `data.frame` of measured environmental properties if applicable.}
  \item{\code{object$spa$rast$his}}{Historical `SpatRaster` of numeric MODIS metrics and other vars if applicable.}
  \item{\code{object$*}}{Spatial and biogeographic traits if applicable.}
}
}
\description{
This function is implemented to download numeric MODIS metrics from `https://cmr.earthdata.nasa.gov`.
}
\examples{
# Create a microgeo dataset
data(qtp)
showtext::showtext_auto(enable = TRUE)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()

# Download numeric MODIS metrics of research region
dataset.dts \%<>\% get_modis_num_metrics(username = "username", password = "password",
                                       prod.typ = "Terra", out.dir = "test/microgeo_data")
dataset.dts \%>\% show_dataset()

# Visualize NDVI
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$his$NDVI,
                  color = colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100)) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()

# Visualize EVI
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$his$EVI,
                  color = colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100)) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()
}
\seealso{
\code{\link[microgeo:show_modis_num_metrics]{microgeo::show_modis_num_metrics()}}
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_spatraster]{microgeo::add_spatraster()}}
\code{\link[microgeo:add_label]{microgeo::add_label()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
\code{\link[microgeo:add_crs]{microgeo::add_crs()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
