% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-spatial-trait.R
\name{get_soilcn}
\alias{get_soilcn}
\title{Process soil metrics of CHINA}
\usage{
get_soilcn(
  dataset,
  measures = c("PH", "SOM", "TN", "TP", "TK", "AN", "AP", "AK", "CEC", "H", "AL", "CA",
    "MG", "K", "NA", "SA", "SI", "CL", "GRAV", "BD", "POR"),
  depth = c(0.045, 0.091, 0.166, 0.289, 0.493, 0.829, 1.383, 2.296),
  out.dir = "microgeo_data"
)
}
\arguments{
\item{dataset}{A microgeo dataset with the class of `MicrogeoDataset`.}

\item{measures}{Variables to be processed. Default is all avaliable variables.}

\item{depth}{Soil depth. It should be one of `0.045, 0.091, 0.166, 0.289, 0.493, 0.829, 1.383, 2.296`, which means the
depth ranges of 0-0.045, 0.045-0.091, 0.091-0.166, 0.166-0.289, 0.289-0.493, 0.493-0.829, 0.829-1.383 & 1.383-2.296 m.
Default is `0.045`.}

\item{out.dir}{Directory path to save downloaded files. Default is `microgeo_data`}
}
\value{
A `MicrogeoDataset` class with the following components:
\describe{
  \item{\code{object$mat}}{A `data.frame` of ASV/gene abundance.}
  \item{\code{object$ant}}{A `data.frame` of ASV/gene anootations.}
  \item{\code{object$met}}{A `data.frame` of sample information.}
  \item{\code{object$map}}{A `SpatialPolygonsDataFrame` of map.}
  \item{\code{object$phy}}{A phylogenetic tree with `newick` format if applicable.}
  \item{\code{object$env}}{A `data.frame` of measured environmental properties if applicable.}
  \item{\code{object$spa$rast$his}}{Historical `SpatRaster` of soil metrics (CHINA) and other vars if applicable.}
  \item{\code{object$*}}{Spatial and biogeographic traits if applicable.}
}
}
\description{
This function could process soil metrics downloaded from `http://globalchange.bnu.edu.cn/research/soil2`.
Because of the limitations in copyrights, the spatial data of China soil properties should be manually downloaded from
\href{http://globalchange.bnu.edu.cn/research/soil2}{here}.
}
\examples{
# Create a microgeo dataset
data(qtp)
showtext::showtext_auto(enable = TRUE)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()

# Process soil metrics of CHINA for a research region
dataset.dts \%<>\% get_soilcn(depth = 0.045, out.dir = "test/microgeo_data")
dataset.dts \%>\% show_dataset()

# Visualize PH
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$his$PH,
                  color = colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100)) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()

# Visualize SOM
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$his$SOM,
                  color = colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100)) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()

# Visualize TN
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$his$TN,
                  color = colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100)) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()

# Visualize AP
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$his$AP,
                  color = colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100)) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_spatraster]{microgeo::add_spatraster()}}
\code{\link[microgeo:add_label]{microgeo::add_label()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
\code{\link[microgeo:add_crs]{microgeo::add_crs()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
