% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-spatial-trait.R
\name{get_soilgrid}
\alias{get_soilgrid}
\title{Download soil metrics from SoilGRIDS}
\usage{
get_soilgrid(
  dataset,
  measures = c("bdod", "cfvo", "nitrogen", "phh2o", "sand", "silt", "soc", "ocd"),
  depth = c(5, 15, 30, 60, 100, 200),
  out.dir = "microgeo_data"
)
}
\arguments{
\item{dataset}{A microgeo dataset with the class of `MicrogeoDataset`.}

\item{measures}{Variables to be downloaded. Default is all avaliable variables. See \code{geodata::soil_world()}.}

\item{depth}{Soil depth. One of `c(5, 15, 30, 60, 100, 200)`, which means the depth ranges of 0-5, 5-15, 15-30, 30-60,
60-100, 100-200 cm. Default is `5`.}

\item{out.dir}{Directory path to save downloaded files. Default is `microgeo_data`}
}
\value{
A `MicrogeoDataset` class with the following components:
\describe{
  \item{\code{object$mat}}{A `data.frame` of ASV/gene abundance.}
  \item{\code{object$ant}}{A `data.frame` of ASV/gene anootations.}
  \item{\code{object$met}}{A `data.frame` of sample information.}
  \item{\code{object$map}}{A `SpatialPolygonsDataFrame` of map.}
  \item{\code{object$phy}}{A phylogenetic tree with `newick` format if applicable.}
  \item{\code{object$env}}{A `data.frame` of measured environmental properties if applicable.}
  \item{\code{object$spa$rast$his}}{Historical `SpatRaster` of soil metrics (SoilGRIDS) and other vars if applicable.}
  \item{\code{object$*}}{Spatial and biogeographic traits if applicable.}
}
}
\description{
This function is used to download soil metrics from SoilGRIDS (`https://www.isric.org/explore/soilgrids`)
by using \code{geodata::soil_world()}.
}
\examples{
# Create a microgeo dataset
data(qtp)
showtext::showtext_auto(enable = TRUE)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()

# Download soil metrics from the SoilGRIDS for the research region
dataset.dts \%<>\% get_soilgrid(depth = 5, out.dir = "test/microgeo_data")
dataset.dts \%>\% show_dataset()

# Visualize phh2o
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$his$phh2o,
                  color = colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100)) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()

# Visualize soc
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$his$soc,
                  color = colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100)) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()

# Visualize nitrogen
dataset.dts$map \%>\% plot_bmap() \%>\%
   add_spatraster(spat.raster = dataset.dts$spa$rast$his$nitrogen,
                  color = colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100)) \%>\%
   add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
   add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()
}
\seealso{
\code{\link[geodata:soil_world]{geodata::soil_world()}}
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_spatraster]{microgeo::add_spatraster()}}
\code{\link[microgeo:add_label]{microgeo::add_label()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
\code{\link[microgeo:add_crs]{microgeo::add_crs()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
