% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-operate.R
\name{grid_map}
\alias{grid_map}
\title{Add grids to a map}
\usage{
grid_map(map, res = 1.5)
}
\arguments{
\item{map}{A geographic map with the class of `SpatialPolygonsDataFrame`.}

\item{res}{Resolution of grids (minutes of a degree). A larger value would generate larger cells. Default is `1.5`.}
}
\value{
A `SpatialPolygonsDataFrame`.
}
\description{
This function is used to add grids to a map based on a specified spatial resolution.
}
\examples{
##### Example 1: add grids on the research areas subsetted from China map (`DataV.GeoAtlas`) #####
map <- read_aliyun_map(adcode = c("540000", "630000", "510000"))
head(map@data)
map \%>\% plot_bmap()
map \%<>\% grid_map(res = 1.5)
head(map@data)
map \%>\% plot_bmap()

##### Example 2: add grids on the map of Australia #####
map <- read_gadm_map(iso = 'Australia', out.dir = 'test')
map \%<>\% trans_map_fmt(var = 'NAME_1')
head(map@data)
map \%>\% plot_bmap()
map \%<>\% grid_map(res = 1.5)
head(map@data)
map \%>\% plot_bmap()

##### Example 3: add grids on the research areas subsetted from China map (`ESRI Shapefile`) #####
map <- system.file("shapefiles/china-map", "china.shp", package = "microgeo") \%>\% read_shp_map()
map \%<>\% terra::subset(map$FENAME \%in\% c("Xizang Zizhiqu", "Qinghai Sheng", "Sichuan Sheng"))
map \%<>\% trans_map_fmt(var = 'FCNAME')
head(map@data)
map \%>\% plot_bmap()
map \%<>\% grid_map(res = 1.5)
head(map@data)
map \%>\% plot_bmap()

##### Example 4: add grids on the research areas that are directly loaded from a `ESRI Shapefile` #####
map <- system.file("shapefiles/qtp-map", "DBATP_Polygon.shp", package = "microgeo") \%>\% read_shp_map()
map@data$NAME <- "Qinghai-Tibet Plateau" # Add a column to meet the demands of `microgeo::trans_map_fmt()`
map \%<>\% trans_map_fmt(var = 'NAME')
head(map@data)
map \%>\% plot_bmap()
map \%<>\% grid_map(res = 1.5)
head(map@data)
map \%>\% plot_bmap()
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:read_gadm_map]{microgeo::read_gadm_map()}}
\code{\link[microgeo:read_shp_map]{microgeo::read_shp_map()}}
\code{\link[microgeo:trans_map_fmt]{microgeo::trans_map_fmt()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
