% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-interp-model.R
\name{interp_kri}
\alias{interp_kri}
\title{Kriging interpolation}
\usage{
interp_kri(
  map,
  met,
  dat,
  var,
  model = c("Mat", "Exp", "Sph", "Gau"),
  n = 4000,
  trim.dup = FALSE,
  test.model = FALSE,
  ...
)
}
\arguments{
\item{map}{A `SpatialPolygonsDataFrame` of geographic map.}

\item{met}{A `data.frame` of sample information. Row names should be sample ids and the column names must be variables
(e.g., `longitude`, `latitude` and `group`). Longitude and latitude must be included in this `data.frame` as we mainly
focus on the spatial patterns of microbial traits.}

\item{dat}{A `data.frame` containing the variables to be interpolated.}

\item{var}{Name of variable to be interpolated.}

\item{model}{Model for kriging interpolation. e.g., `Mat`, `Exp`, `Sph` and `Gau`. Calling \code{gstat::vgm()} without
a argument returns a `data.frame` with available models. See \code{gstat::vgm()} for more details. Default is `Mat`.}

\item{n}{Approximate sample size. A large value means a high resolution. Default is `4000`. See \code{sp::spsample()}.}

\item{trim.dup}{Whether to randomly remove sampling sites with duplicated coordinates? Default is `FALSE`.}

\item{test.model}{Whether to test selected model (<model>) rather than directly perform kriging interpolation? Default
is `FALSE`.}

\item{...}{Parameters parsed by \code{gstat::krige()}.}
}
\value{
A `SpatRaster` if the <test.model> is `FALSE`.
}
\description{
This function is implemented to perform the kriging interpolation.
}
\examples{
# Create a microgeo dataset
data(qtp)
showtext::showtext_auto(enable = TRUE)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()

# Tidy microgeo dataset, and calculate microbial alpha diversity indices
dataset.dts \%<>\% rarefy_count_table()
dataset.dts \%<>\% tidy_dataset()
dataset.dts \%<>\% calc_alpha_div(measures = c("observed", "shannon"))
dataset.dts \%>\% show_dataset()

# Test the `Sph` and `Mat` models
interp_kri(map = dataset.dts$map, met = dataset.dts$met, dat = dataset.dts$div$alpha,
           var = 'observed', model = 'Sph', test.model = TRUE) # error
interp_kri(map = dataset.dts$map, met = dataset.dts$met, dat = dataset.dts$div$alpha,
           var = 'observed', model = 'Sph', test.model = TRUE, trim.dup = TRUE)
interp_kri(map = dataset.dts$map, met = dataset.dts$met, dat = dataset.dts$div$alpha,
           var = 'shannon', model = 'Mat', test.model = TRUE) # error
interp_kri(map = dataset.dts$map, met = dataset.dts$met, dat = dataset.dts$div$alpha,
           var = 'shannon', model = 'Mat', test.model = TRUE, trim.dup = TRUE)

# Perform kriging interpolation for alpha diversity indices
kri.rst.observed <- interp_kri(map = dataset.dts$map, met = dataset.dts$met,
                               dat = dataset.dts$div$alpha, var = 'observed',
                               model = 'Sph', trim.dup = TRUE)
kri.rst.shannon <- interp_kri(map = dataset.dts$map, met = dataset.dts$met,
                               dat = dataset.dts$div$alpha, var = 'shannon',
                               model = 'Mat', trim.dup = TRUE)

# Visualize the interpolated results
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_spatraster(spat.raster = kri.rst.observed) \%>\%
    add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = "X.CENTER", lat.var = "Y.CENTER") \%>\%
    add_scale_bar() \%>\% add_north_arrow() \%>\% add_crs()
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_spatraster(spat.raster = kri.rst.shannon) \%>\%
    add_label(dat = dataset.dts$map@data, lab.var = 'NAME', lon.var = "X.CENTER", lat.var = "Y.CENTER") \%>\%
    add_scale_bar() \%>\% add_north_arrow() \%>\% add_crs()
}
\seealso{
\code{\link[sp:spsample]{sp::spsample()}}
\code{\link[gstat:vgm]{gstat::vgm()}}
\code{\link[gstat:fit.variogram]{gstat::fit.variogram()}}
\code{\link[gstat:krige]{gstat::krige()}}
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:rarefy_count_table]{microgeo::rarefy_count_table()}}
\code{\link[microgeo:tidy_dataset]{microgeo::tidy_dataset()}}
\code{\link[microgeo:calc_alpha_div]{microgeo::calc_alpha_div()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_spatraster]{microgeo::add_spatraster()}}
\code{\link[microgeo:add_label]{microgeo::add_label()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
\code{\link[microgeo:add_crs]{microgeo::add_crs()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
