% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-interp-model.R
\name{interp_nen}
\alias{interp_nen}
\title{Nearest neighbour interpolation}
\usage{
interp_nen(map, met, dat, var, trim.dup = FALSE)
}
\arguments{
\item{map}{A `SpatialPolygonsDataFrame` of geographic map.}

\item{met}{A `data.frame` of sample information. Row names should be sample ids and the column names must be variables
(e.g., `longitude`, `latitude` and `group`). Longitude and latitude must be included in this `data.frame` as we mainly
focus on the spatial patterns of microbial traits.}

\item{dat}{A `data.frame` containing the variables to be interpolated.}

\item{var}{Name of variable to be interpolated.}

\item{trim.dup}{Whether to randomly remove sampling sites with duplicated coordinates? Default is `FALSE`.}
}
\value{
An R `list` with the following components:
\describe{
  \item{\code{object$res}}{Data for ggplot2 visualization.}
  \item{\code{object$var}}{Name of interpolated variable.}
  \item{\code{object$type}}{Type of interpolation, which is `'nen'`.}
}
}
\description{
This function is used to perform nearest neighbour interpolation (Voronoi diagram).
}
\examples{
# Create a microgeo dataset
data(qtp)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()

# Tidy microgeo dataset, and calculate microbial alpha diversity indices
dataset.dts \%<>\% rarefy_count_table()
dataset.dts \%<>\% tidy_dataset()
dataset.dts \%<>\% calc_alpha_div(measures = c("observed", "shannon"))
dataset.dts \%>\% show_dataset()

# Nearest neighbour interpolation for alpha diversity indices
observed.rst <- interp_nen(map = dataset.dts$map, met = dataset.dts$met,
                           dat = dataset.dts$div$alpha, var = 'observed', trim.dup = TRUE)
shannon.rst <- interp_nen(map = dataset.dts$map, met = dataset.dts$met,
                          dat = dataset.dts$div$alpha, var = 'shannon', trim.dup = TRUE)

# Visualize the interpolated results
observed.rst \%>\% plot_nmap() \%>\% add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()
shannon.rst \%>\% plot_nmap() \%>\% add_north_arrow() \%>\% add_scale_bar() \%>\% add_crs()
}
\seealso{
\code{\link[sf:st_voronoi]{sf::st_voronoi()}}
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:rarefy_count_table]{microgeo::rarefy_count_table()}}
\code{\link[microgeo:tidy_dataset]{microgeo::tidy_dataset()}}
\code{\link[microgeo:calc_alpha_div]{microgeo::calc_alpha_div()}}
\code{\link[microgeo:plot_nmap]{microgeo::plot_nmap()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
\code{\link[microgeo:add_crs]{microgeo::add_crs()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
