% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-tools.R
\name{mask_spatraster_by_cla}
\alias{mask_spatraster_by_cla}
\title{Mask a SpatRaster through using another SpatRaster as a reference}
\usage{
mask_spatraster_by_cla(tar.spat, ref.spat, use.class)
}
\arguments{
\item{tar.spat}{Target `SpatRaster`.}

\item{ref.spat}{Reference `SpatRaster`.}

\item{use.class}{Class number. See \href{https://lpdaac.usgs.gov/documents/1409/MCD12_User_Guide_V61.pdf}{here}.}
}
\value{
A `SpatRaster`.
}
\description{
This function is used to mask a SpatRaster using another SpatRaster with classification values as a ref.
}
\examples{
# Create a microgeo dataset
data(qtp)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()
dataset.dts \%<>\% get_modis_cla_metrics(username = "username", password = "passwd", out.dir = "test/microgeo_data")
dataset.dts \%>\% show_dataset()
dataset.dts \%<>\% rarefy_count_table()
dataset.dts \%<>\% tidy_dataset()
dataset.dts \%<>\% calc_alpha_div(measures = c("observed", "shannon"))
dataset.dts \%>\% show_dataset()

# Perform kriging interpolation for alpha diversity indices
kri.rst.shannon <- interp_kri(map = dataset.dts$map,
                              met = dataset.dts$met,
                              dat = dataset.dts$div$alpha,
                              var = 'shannon', model = 'Mat', trim.dup = TRUE)

# Mask the results by using Grasslands(10) and Barren(16)
kri.rst.shannon.masked <- mask_spatraster_by_cla(tar.spat = kri.rst.shannon,
                                                 ref.spat = dataset.dts$spa$rast$cla$LC_Type1,
                                                 use.class = c(10, 16))
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_spatraster(spat.raster = kri.rst.shannon.masked) \%>\%
    add_scale_bar() \%>\% add_north_arrow() \%>\% add_crs()

# Mask the results by using Grasslands(10)
kri.rst.shannon.masked <- mask_spatraster_by_cla(tar.spat = kri.rst.shannon,
                                                 ref.spat = dataset.dts$spa$rast$cla$LC_Type1,
                                                 use.class = 10)
dataset.dts$map \%>\% plot_bmap() \%>\%
    add_spatraster(spat.raster = kri.rst.shannon.masked) \%>\%
    add_scale_bar() \%>\% add_north_arrow() \%>\% add_crs()
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:get_modis_cla_metrics]{microgeo::get_modis_cla_metrics()}}
\code{\link[microgeo:rarefy_count_table]{microgeo::rarefy_count_table()}}
\code{\link[microgeo:tidy_dataset]{microgeo::tidy_dataset()}}
\code{\link[microgeo:calc_alpha_div]{microgeo::calc_alpha_div()}}
\code{\link[microgeo:interp_kri]{microgeo::interp_kri()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_spatraster]{microgeo::add_spatraster()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
\code{\link[microgeo:add_crs]{microgeo::add_crs()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
