% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-operate.R
\name{merge_dfs_to_map}
\alias{merge_dfs_to_map}
\title{Merge a `data.frame` with a map}
\usage{
merge_dfs_to_map(map, dat, met, med = c("mean", "median"))
}
\arguments{
\item{map}{Geographic map with a class of `SpatialPolygonsDataFrame`.}

\item{dat}{A `data.frame` containing the variables to be merged with the <map>. Such a `data.frame` should not include
sample names (please set sample names as the row names). You can include any number of variables in this `data.frame`.}

\item{met}{A `data.frame` of sample information. Row names should be sample ids and the column names must be variables
(e.g., `longitude`, `latitude` and `group`). Longitude and latitude must be included in this `data.frame` as we mainly
focus on the spatial patterns of microbial traits.}

\item{med}{Method to aggregate numeric variables. Select one from `mean` and `median`. Default is `mean`.}
}
\value{
A `SpatialPolygonsDataFrame`.
}
\description{
This function is used to merge a `data.frame` with a map. Values in this `data.frame` must be numeric.
}
\examples{
# Create a standard microgeo dataset
data(qtp)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%<>\% rarefy_count_table()
dataset.dts \%<>\% tidy_dataset()
dataset.dts \%<>\% calc_alpha_div(measures = c("observed", "shannon"))
dataset.dts \%>\% show_dataset()

# Merge alpha diversity indices to a common map
common.map.mean <- merge_dfs_to_map(map = dataset.dts$map, dat = dataset.dts$div$alpha,
                                    met = dataset.dts$met, med = 'mean')
common.map.median <- merge_dfs_to_map(map = dataset.dts$map, dat = dataset.dts$div$alpha,
                                    met = dataset.dts$met, med = 'median')
View(common.map.mean@data)
View(common.map.median@data)

# Merge alpha diversity indices to a gridded map
gridded.map <- dataset.dts$map \%>\% grid_map(res = 1.5)
gridded.map.mean <- merge_dfs_to_map(map = gridded.map, dat = dataset.dts$div$alpha,
                                     met = dataset.dts$met, med = 'mean')
gridded.map.median <- merge_dfs_to_map(map = gridded.map, dat = dataset.dts$div$alpha,
                                     met = dataset.dts$met, med = 'median')
View(gridded.map.mean@data)
View(gridded.map.median@data)
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:rarefy_count_table]{microgeo::rarefy_count_table()}}
\code{\link[microgeo:tidy_dataset]{microgeo::tidy_dataset()}}
\code{\link[microgeo:calc_alpha_div]{microgeo::calc_alpha_div()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:grid_map]{microgeo::grid_map()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
