% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-operate.R
\name{merge_mtx_to_map}
\alias{merge_mtx_to_map}
\title{Merge a distance `matrix` with a map}
\usage{
merge_mtx_to_map(
  map,
  dat,
  met,
  var = NULL,
  med = c("mean", "median", "dtc.mean", "dtc.median")
)
}
\arguments{
\item{map}{Geographic map with a class of `SpatialPolygonsDataFrame`.}

\item{dat}{A symmetric distance matrix.}

\item{met}{A `data.frame` of sample information. Row names should be sample ids and the column names must be variables
(e.g., `longitude`, `latitude` and `group`). Longitude and latitude must be included in this `data.frame` as we mainly
focus on the spatial patterns of microbial traits.}

\item{var}{A string to name the variable merged with a map. If it is `NULL`, the default, we will use the `dist_*`.}

\item{med}{Method to aggregate the variable. Select one from `mean`, `median`, `dtc.mean` (mean distances to centroid)
and `dtc.median` (median distances to centroid). Default is `mean`.}
}
\value{
A `SpatialPolygonsDataFrame`.
}
\description{
This function is used to merge a distance matrix with a map. Values in this `matrix` must be numeric.
}
\examples{
# Create a standard microgeo dataset
data(qtp)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%<>\% rarefy_count_table()
dataset.dts \%<>\% tidy_dataset()
dataset.dts \%<>\% calc_beta_div(measures = c("bray", "jaccard"))
dataset.dts \%>\% show_dataset()

# Merge distance matrix to a common map
map.mean <- merge_mtx_to_map(map = dataset.dts$map, dat = dataset.dts$div$beta$bray,
                             met = dataset.dts$met, var = 'bray', med = 'mean')
map.median <- merge_mtx_to_map(map = dataset.dts$map, dat = dataset.dts$div$beta$bray,
                               met = dataset.dts$met, var = 'bray', med = 'median')
map.dtc.mean <- merge_mtx_to_map(map = dataset.dts$map, dat = dataset.dts$div$beta$bray,
                                 met = dataset.dts$met, var = 'bray', med = 'dtc.mean')
map.dtc.median <- merge_mtx_to_map(map = dataset.dts$map, dat = dataset.dts$div$beta$bray,
                                 met = dataset.dts$met, var = 'bray', med = 'dtc.median')
View(map.mean@data)
View(map.median@data)
View(map.dtc.mean@data)
View(map.dtc.median@data)

# Merge distance matrix to a gridded map
gridded.map <- dataset.dts$map \%>\% grid_map(res = 1.5)
gridded.map.mean <- merge_mtx_to_map(map = gridded.map, dat = dataset.dts$div$beta$bray,
                                     met = dataset.dts$met, var = 'bray', med = 'mean')
gridded.map.median <- merge_mtx_to_map(map = gridded.map, dat = dataset.dts$div$beta$bray,
                                       met = dataset.dts$met, var = 'bray', med = 'median')
gridded.map.dtc.mean <- merge_mtx_to_map(map = gridded.map, dat = dataset.dts$div$beta$bray,
                                         met = dataset.dts$met, var = 'bray', med = 'dtc.mean')
gridded.map.dtc.median <- merge_mtx_to_map(map = gridded.map, dat = dataset.dts$div$beta$bray,
                                           met = dataset.dts$met, var = 'bray', med = 'dtc.median')
View(gridded.map.mean@data)
View(gridded.map.median@data)
View(gridded.map.dtc.mean@data)
View(gridded.map.dtc.median@data)
}
\seealso{
\code{\link[vegan:betadisper]{vegan:betadisper()}}
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:rarefy_count_table]{microgeo::rarefy_count_table()}}
\code{\link[microgeo:tidy_dataset]{microgeo::tidy_dataset()}}
\code{\link[microgeo:calc_beta_div]{microgeo::calc_beta_div()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:grid_map]{microgeo::grid_map()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
