% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-visualize.R
\name{plot_bmap}
\alias{plot_bmap}
\title{Plot a basic map of the research area}
\usage{
plot_bmap(
  map,
  var = NULL,
  ord = NULL,
  fill = NA,
  color = "gray30",
  width = 0.8,
  alpha = 0.8,
  gg.theme = theme_bw(),
  bg.color = "white",
  gd.color = "gray90",
  axis.text.size = 12,
  legend.position = c("right", "left", "top", "bottom"),
  legend.title.size = 14,
  legend.text.size = 12,
  show.legend.rect = FALSE
)
}
\arguments{
\item{map}{Geographic map with a class of `SpatialPolygonsDataFrame`.}

\item{var}{Variable used to fill map area. If there is only one polygon in the <map>, the value of this argument would
be ignored and the map area would be filled up by the <fill> (only one color value) or the first element of the <fill>
if there are mutiple color values in the <fill>. Default is `NULL`.}

\item{ord}{Character vector for the display order if the <var> is a `character` or `factor`. Default is `NULL`.}

\item{fill}{A color value of a color vector used to fill the map region. This argument should not be `NA` if the <var>
is not `NULL` (You can set the <fill> as `'auto'`). If the <var> is a `number`, it is better to set a continuous color
vector for the <fill>, for example, `colorRampPalette(RColorBrewer::brewer.pal(11, 'RdYlGn'))(1000) %>% rev()`. If the
<var> is a `character` or `factor`, it is better to set a discrete color vector for the <fill>, e.g., the color vector
could be generated by `randomcoloR::randomColor(count = 1000)`. If the <fill> is `'auto'` and the <var> is not `NULL`,
the colors would be produced automatically based on the above described schema. Default is `NA`.}

\item{color}{Line color of border. Default is `gray30`.}

\item{width}{Line width of border. Default is `0.8`.}

\item{alpha}{Transparency of fill color. Default is `0.8`.}

\item{gg.theme}{The ggplot2 theme. Default is `theme_bw()`.}

\item{bg.color}{Color of panel background. Default is `white`.}

\item{gd.color}{Color of panel grid major. Default is `gray90`.}

\item{axis.text.size}{Size of axis text. Default is `12`.}

\item{legend.position}{Legend position. Select a value from `'right', 'left', 'top', 'bottom'` or set a numeric vector
like `c(0.1, 0.2)`. Default is `right`.}

\item{legend.title.size}{Text size of legend title. Default is `14`.}

\item{legend.text.size}{Text size of legend label. Default is `12`.}

\item{show.legend.rect}{Should the rectangle around the legend be displayed? Default is `FALSE`.}
}
\value{
A ggplot2 object.
}
\description{
This function is used to plot a basic map of the research area. All data for the visualizations should be
included in the <map> (a `SpatialPolygonsDataFrame` class).
}
\examples{
##### Initialization of the visualizations #####
data(common.map.mean)
data(gridded.map.mean)
showtext::showtext_auto(enable = TRUE)

##### Example 1: plot maps without any fill color #####
common.map.mean  \%>\% plot_bmap()
gridded.map.mean \%>\% plot_bmap()

##### Example 2: plot maps filled by gray color #####
common.map.mean  \%>\% plot_bmap(fill = 'gray50', alpha = 0.5)
common.map.mean  \%>\% plot_bmap(fill = 'gray50', color = 'blue', alpha = 0.5)
gridded.map.mean \%>\% plot_bmap(fill = 'gray50', color = 'darkgreen', alpha = 0.5)

##### Example 3: plot maps automatically filled by the colors defined by a variable #####
common.map.mean  \%>\% plot_bmap(var = 'NAME', alpha = 0.5, fill = 'auto')
gridded.map.mean \%>\% plot_bmap(var = 'sample.num', fill = 'auto')
gridded.map.mean \%>\% plot_bmap(var = 'shannon_mean', fill = 'auto')
gridded.map.mean \%>\% plot_bmap(var = 'NAME', alpha = 0.5, fill = 'auto', legend.position = 'none')

##### Example 4: plot maps automatically filled by custom colors #####
common.map.mean  \%>\% plot_bmap(var = 'NAME', ord = c('西藏自治区', '青海省', '四川省'),
                               fill = RColorBrewer::brewer.pal(9, "Set1")[1:3])
common.map.mean  \%>\% plot_bmap(var = 'NAME', ord = c('四川省', '青海省', '西藏自治区'),
                               fill = RColorBrewer::brewer.pal(9, "Set1")[4:6])
gridded.map.mean \%>\% plot_bmap(var = 'NAME', legend.position = 'none',
                               fill = colorRampPalette(RColorBrewer::brewer.pal(11, 'BrBG'))(133) \%>\% rev)
gridded.map.mean \%>\% plot_bmap(var = 'sample.num',
                               fill = colorRampPalette(RColorBrewer::brewer.pal(11, 'BrBG'))(100) \%>\% rev)
gridded.map.mean \%>\% plot_bmap(var = 'shannon_mean',
                               fill = colorRampPalette(RColorBrewer::brewer.pal(11, 'BrBG'))(100) \%>\% rev)
gridded.map.mean \%>\% plot_bmap(var = 'shannon_mean', show.legend.rect = TRUE,
                               fill = colorRampPalette(RColorBrewer::brewer.pal(11, 'BrBG'))(100) \%>\% rev)
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
