% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-visualize.R
\name{plot_nmap}
\alias{plot_nmap}
\title{Plot a map for nearest neighbour interpolation}
\usage{
plot_nmap(
  dat,
  fill = colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlGn"))(100) \%>\% rev,
  color = "gray50",
  alpha = 0.8,
  width = 0.8,
  size = 12,
  gg.theme = theme_bw(),
  bg.color = "white",
  gd.color = "gray80",
  legend.position = c("right", "left", "top", "bottom"),
  legend.title.size = 14,
  legend.text.size = 12,
  show.legend.rect = FALSE
)
}
\arguments{
\item{dat}{An R list returned by \code{microgeo::interp_nen()}.}

\item{fill}{Fill color of map area. Default is `rev(colorRampPalette(RColorBrewer::brewer.pal(11, 'RdYlGn'))(100))`.}

\item{color}{Line color of map border. Default is `gray50`.}

\item{alpha}{Transparency of fill color. Default is `0.8`.}

\item{width}{Line width of map border. Default is `0.8`.}

\item{size}{Size of axis text. Default is `12`.}

\item{gg.theme}{The ggplot2 theme. Default is `theme_bw()`.}

\item{bg.color}{Color of panel background. Default is `white`.}

\item{gd.color}{Color of panel grid major. Default is `gray80`.}

\item{legend.position}{Legend position. Select a value from `'right', 'left', 'top', 'bottom'` or set a numeric vector
like `c(0.1, 0.2)`. Default is `right`.}

\item{legend.title.size}{Text size of legend title. Default is `14`.}

\item{legend.text.size}{Text size of legend label. Default is `12`.}

\item{show.legend.rect}{Should the rectangle around the legend be displayed? Default is `FALSE`.}
}
\value{
A ggplot2 object.
}
\description{
This function is designed to visualize the results of nearest neighbour interpolation.
}
\examples{
data(nen.interp)
nen.interp$shannon  \%>\% plot_nmap()
nen.interp$observed \%>\% plot_nmap()
nen.interp$observed \%>\% plot_nmap(show.legend.rect = TRUE)
nen.interp$observed \%>\% plot_nmap(fill = rev(colorRampPalette(RColorBrewer::brewer.pal(11, 'Spectral'))(100)))
}
\seealso{
\code{\link[microgeo:interp_nen]{microgeo::interp_nen()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
