% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-operate.R
\name{read_aliyun_map}
\alias{read_aliyun_map}
\title{Read one or multiple areas of China map from the DataV.GeoAtlas (Aliyun, China)}
\usage{
read_aliyun_map(
  adcode = c(540000, 630000, 510000),
  sub.area = FALSE,
  crs.string = "EPSG:4326"
)
}
\arguments{
\item{adcode}{Adcode of the map area(s). For example, the `650000` means Xinjiang, and the `c(540000, 630000, 510000)`
represents Xizang, Qinghai, and Sichuan. You can visit the `http://datav.aliyun.com/portal/school/atlas/area_selector`
to select one or mutiple adcodes for your interested areas. Default is `c(540000, 630000, 510000)`.}

\item{sub.area}{Whether to include sub-layers? Default is `FALSE`.}

\item{crs.string}{Coordinate Reference System (CRS). Default is `"EPSG:4326"`.}
}
\value{
A `SpatialPolygonsDataFrame`.
}
\description{
This function is implemented to download one or multiple area(s) of China map from Aliyun DataV.GeoAtlas.
Undoubtedly, it is user friendly for Chinese since the boundary data is relatively accurate and is very likely to have
undergone official review. It is a better choice to use \code{microgeo::read_aliyun_map()} when you publishing several
graphics in official publications if you are unsure whether your map dataset is accurate.
}
\examples{
## Example 1: using one adcode
map1 <- read_aliyun_map(adcode = 650000)
map2 <- read_aliyun_map(adcode = 650000, sub.area = TRUE)
head(map1@data)
head(map2@data)
map1 \%>\% plot_bmap()
map2 \%>\% plot_bmap()

## Example 1: using mutiple adcodes
map1 <- read_aliyun_map(adcode = c(540000, 630000, 510000))
map2 <- read_aliyun_map(adcode = c(540000, 630000, 510000), sub.area = TRUE)
head(map1@data)
head(map2@data)
map1 \%>\% plot_bmap()
map2 \%>\% plot_bmap()
}
\seealso{
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
