% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-operate.R
\name{read_gadm_map}
\alias{read_gadm_map}
\title{Read a geographic map from the GADM (Global Administrative Areas)}
\usage{
read_gadm_map(
  iso,
  res = 1,
  level = 1,
  version = 3.6,
  out.dir = "GADM",
  crs.string = "EPSG:4326",
  ...
)
}
\arguments{
\item{iso}{Three-letter ISO code or a full country name. Multiple names would be all downloaded and rbinded.}

\item{res}{An integer indicating the level of detail. Only for the version 4.1 of `GADM`. It must be either `1` (high)
or `2` (low). Default is `1`.}

\item{level}{The level of administrative subdivision requested. (starting with `0` for country, then `1` for the first
level of subdivision). Default is `1`.}

\item{version}{Either `latest` or GADM version number (can be `3.6`, `4.0` or `4.1`). Default is `latest`.}

\item{out.dir}{Path for storing the downloaded data. Default is `GADM`.}

\item{crs.string}{Coordinate Reference System (CRS). Default is `"EPSG:4326"`.}

\item{...}{Parameters parsed by \code{geodata::gadm()}.}
}
\value{
A `SpatialPolygonsDataFrame`.
}
\description{
This function is designed to read a map from the GADM (`https://gadm.org/`) using \code{geodata::gadm()}.
Particularly, the Chinese territorial claims are not reflected in the boundary data provided by the GADM, and the data
for provincial/city/district boundaries may not necessarily be the most up-to-date versions. Thus, additional cautions
should be exercised when publishing graphics that utilize GADM data in official publications. The author of microgeo R
package is not responsible for any political disputes and instability caused by the use of this R package by anyone.
}
\examples{
## Example 1: the map of Australia
map <- read_gadm_map(iso = 'Australia', out.dir = 'test')
map \%<>\% trans_map_fmt(var = 'NAME_1')
head(map@data)
map \%>\% plot_bmap()

## Example 2: the map of Washington State, USA
map <- read_gadm_map(iso = 'USA', out.dir = 'test')
map \%<>\% terra::subset(map$NAME_1 == 'Washington')
map \%<>\% trans_map_fmt(var = 'NAME_1')
head(map@data)
map \%>\% plot_bmap()
}
\seealso{
\code{\link[geodata:gadm]{geodata::gadm()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:trans_map_fmt]{microgeo::trans_map_fmt()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
