% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-operate.R
\name{read_shp_map}
\alias{read_shp_map}
\title{Read a geographic map from the ESRI Shapefile}
\usage{
read_shp_map(shpfile, crs.string = "EPSG:4326", ...)
}
\arguments{
\item{shpfile}{A file path of ESRI Shapefile.}

\item{crs.string}{Coordinate Reference System (CRS). Default is `"EPSG:4326"`.}

\item{...}{Parameters parsed by \code{terra::vect()}.}
}
\value{
A `SpatialPolygonsDataFrame`.
}
\description{
This function is used to read a geographic map from the ESRI Shapefile via using \code{terra::vect()}.
}
\examples{
##### Example 1: subset the research areas from a ESRI Shapefile of China map #####
map <- system.file("shapefiles/china-map", "china.shp", package = "microgeo") \%>\% read_shp_map()
map \%<>\% terra::subset(map$FENAME \%in\% c("Xizang Zizhiqu", "Qinghai Sheng", "Sichuan Sheng"))
map \%<>\% trans_map_fmt(var = 'FCNAME')
head(map@data)
map \%>\% plot_bmap()

##### Example 2: directly load the research areas from a ESRI Shapefile #####
map <- system.file("shapefiles/qtp-map", "DBATP_Polygon.shp", package = "microgeo") \%>\% read_shp_map()
map@data$NAME <- "Qinghai-Tibet Plateau" # Add a column to meet the demands of `microgeo::trans_map_fmt()`
map \%<>\% trans_map_fmt(var = 'NAME')
head(map@data)
map \%>\% plot_bmap()
}
\seealso{
\code{\link[terra::vect]{terra::vect()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:trans_map_fmt]{microgeo::trans_map_fmt()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
