% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microgeo-dataset.R
\name{show_dataset}
\alias{show_dataset}
\title{Show the summary of a microgeo dataset}
\usage{
show_dataset(dataset)
}
\arguments{
\item{dataset}{A microgeo dataset with the class of `MicrogeoDataset`.}
}
\description{
This function is implemented to show the summary of a microgeo dataset (a `MicrogeoDataset` class).
}
\examples{
data(qtp)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
map \%>\% plot_bmap() \%>\%
    add_label(dat = map@data, lab.var = 'NAME', lon.var = 'X.CENTER', lat.var = 'Y.CENTER') \%>\%
    add_scale_bar() \%>\% add_north_arrow() \%>\% add_crs()
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_label]{microgeo::add_label()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[microgeo:add_crs]{microgeo::add_crs()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
