% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-tools.R
\name{show_modis_cla_metrics}
\alias{show_modis_cla_metrics}
\title{Show those classification MODIS (Moderate-resolution Imaging Spectroradiometer, `https://modis.gsfc.nasa.gov/`)
metrics avaliable in microgeo R package.}
\usage{
show_modis_cla_metrics()
}
\value{
A `data.frame` of classification MODIS metrics.
}
\description{
This function is used to show all classification MODIS metrics avaliable in microgeo R package.
}
\examples{
show_modis_cla_metrics()
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
