% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-tools.R
\name{subset_cla_spatraster}
\alias{subset_cla_spatraster}
\title{Subset the classification SpatRaster.}
\usage{
subset_cla_spatraster(spat.raster, use.class)
}
\arguments{
\item{spat.raster}{A classification SpatRaster returned by \code{microgeo::get_modis_cla_metrics()}.}

\item{use.class}{Class number. See \href{https://lpdaac.usgs.gov/documents/1409/MCD12_User_Guide_V61.pdf}{here}.}
}
\value{
A `SpatRaster`.
}
\description{
This function is designed to subset the SpatRaster returned by \code{microgeo::get_modis_cla_metrics()}.
}
\examples{
# Create a microgeo dataset
data(qtp)
map <- read_aliyun_map(adcode = c(540000, 630000, 510000))
dataset.dts <- create_dataset(mat = qtp$asv, ant = qtp$tax, met = qtp$met, map = map,
                              phy = qtp$tre, env = qtp$env, lon = "longitude", lat = "latitude")
dataset.dts \%>\% show_dataset()
dataset.dts \%<>\% get_modis_cla_metrics(username = "username", password = "passwd", out.dir = "test/microgeo_data")
dataset.dts \%>\% show_dataset()

# Visualize all values of LC_Type1
dataset.dts$map \%>\%
    plot_bmap() \%>\%
    add_spatraster(spat.raster = dataset.dts$spa$rast$cla$LC_Type1,
                   color = c(RColorBrewer::brewer.pal(12, "Set3"), RColorBrewer::brewer.pal(9, "Set1"))) \%>\%
    add_north_arrow() \%>\% add_scale_bar()

# Visualize Grasslands(10) and Barren(16)
gb.spat.raster <- subset_cla_spatraster(spat.raster = dataset.dts$spa$rast$cla$LC_Type1, use.class = c(10, 16))
dataset.dts$map \%>\%
    plot_bmap() \%>\%
    add_spatraster(spat.raster = gb.spat.raster,
                   color = RColorBrewer::brewer.pal(12, "Set3")[1:2]) \%>\%
    add_north_arrow() \%>\% add_scale_bar()

# Visualize Grasslands(10)
g.spat.raster <- subset_cla_spatraster(spat.raster = dataset.dts$spa$rast$cla$LC_Type1, use.class = 10)
dataset.dts$map \%>\%
    plot_bmap() \%>\%
    add_spatraster(spat.raster = g.spat.raster, color = "green") \%>\%
    add_north_arrow() \%>\% add_scale_bar()
}
\seealso{
\code{\link[microgeo:read_aliyun_map]{microgeo::read_aliyun_map()}}
\code{\link[microgeo:create_dataset]{microgeo::create_dataset()}}
\code{\link[microgeo:show_dataset]{microgeo::show_dataset()}}
\code{\link[microgeo:get_modis_cla_metrics]{microgeo::get_modis_cla_metrics()}}
\code{\link[microgeo:plot_bmap]{microgeo::plot_bmap()}}
\code{\link[microgeo:add_spatraster]{microgeo::add_spatraster()}}
\code{\link[microgeo:add_north_arrow]{microgeo::add_north_arrow()}}
\code{\link[microgeo:add_scale_bar]{microgeo::add_scale_bar()}}
\code{\link[microgeo:get_modis_cla_metrics]{microgeo::get_modis_cla_metrics()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
