% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-operate.R
\name{trans_map_fmt}
\alias{trans_map_fmt}
\title{Convert the non-standardized `SpatialPolygonsDataFrame` to a microgeo-compatible one}
\usage{
trans_map_fmt(map, var)
}
\arguments{
\item{map}{A geographic map with the class of `SpatialPolygonsDataFrame`. Only accepts the `SpatialPolygonsDataFrame`
returned by \code{microgeo::read_gadm_map()} and \code{microgeo::read_shp_map()}.}

\item{var}{Which variable in the <map> would be used as the name(s) of research area(s)?}
}
\value{
A `SpatialPolygonsDataFrame`.
}
\description{
This function is designed to transform the `SpatialPolygonsDataFrame` to a microgeo-compatible one. It is
used for a `SpatialPolygonsDataFrame` returned by \code{microgeo::read_gadm_map()} or \code{microgeo::read_shp_map()}.
}
\examples{
## Example 1: the map of Australia
map <- read_gadm_map(iso = 'Australia', out.dir = 'test')
head(map@data)
map \%<>\% trans_map_fmt(var = 'NAME_1')
head(map@data)

## Example 2: the map of Washington State, USA
map <- read_gadm_map(iso = 'USA', out.dir = 'test')
head(map@data)
map \%<>\% terra::subset(map$NAME_1 == 'Washington')
head(map@data)
map \%<>\% trans_map_fmt(var = 'NAME_1')
head(map@data)
}
\seealso{
\code{\link[microgeo:read_gadm_map]{microgeo::read_gadm_map()}}
\code{\link[microgeo:read_shp_map]{microgeo::read_shp_map()}}
}
\author{
Li Chaonan (Ecological Security and Protection Key Laboratory of Sichuan Province, Mianyang Normal University)
}
